### 2021-03-02 #3080 @whedon
[PRE REVIEW]: FitBenchmarking: an open source Python package comparing data fitting software
**Submitting author:** @tyronerees (<a href="http://orcid.org/0000-0003-0476-2259">Tyrone Rees</a>)
**Repository:** <a href="https://github.com/fitbenchmarking/fitbenchmarking" target ="_blank">https://github.com/fitbenchmarking/fitbenchmarking</a>
**Version:** v 0.1.0
