### 2021-03-02 18dd2ca
update minuit_controller to support iminuit 2.x series
