### 2021-03-02 4f2fb76
update minuit_controller to support iminuit 2.x series
(cherry picked from commit 18dd2cae6ab0f3532dd0da6a8afdc4b392981a44)
