+++ fitbenchmarking/controllers/minuit_controller.py
@@ -5,6 +5,7 @@ using the iminuit python interface
 http://iminuit.readthedocs.org
 """
 from iminuit import Minuit
+import iminuit
 import numpy as np
 
 from fitbenchmarking.controllers.base_controller import Controller
@@ -23,6 +24,9 @@ class MinuitController(Controller):
         :type problem: FittingProblem
         """
         super(MinuitController, self).__init__(problem)
+        if int(iminuit.__version__[:1]) < 2:
+            raise Exception("iminuit version " + iminuit.__version__ + 
+                            " is not supported, please upgrade to at least version 2.0.0")
         self._popt = None
         self._initial_step = None
         self._minuit_problem = None
@@ -45,24 +49,24 @@ class MinuitController(Controller):
         self._initial_step = 0.1 * np.array(self.initial_params)
         # set small steps to something sensible(?)
         self._initial_step[self._initial_step < 1e-12] = 1e-12
-        self._minuit_problem = Minuit.from_array_func(self.problem.eval_r_norm,
-                                                      self.initial_params,
-                                                      error=self._initial_step,
-                                                      errordef=1)
+        self._minuit_problem = Minuit(self.problem.eval_r_norm,
+                                      self.initial_params)
+        self._minuit_problem.errordef = 1
+        self._minuit_problem.errors = self._initial_step
 
     def fit(self):
         """
         Run problem with Minuit
         """
         self._minuit_problem.migrad()  # run optimizer
-        self._status = 0 if self._minuit_problem.migrad_ok() else 1
+        self._status = 0 if self._minuit_problem.valid else 1
 
     def cleanup(self):
         """
         Convert the result to a numpy array and populate the variables results
         will be read from
         """
-        fmin = self._minuit_problem.get_fmin()
+        fmin = self._minuit_problem.fmin
         if self._status == 0:
             self.flag = 0
         elif fmin.has_reached_call_limit:
@@ -70,5 +74,5 @@ class MinuitController(Controller):
         else:
             self.flag = 2
 
-        self._popt = self._minuit_problem.np_values()
+        self._popt = np.array(self._minuit_problem.values)
         self.final_params = self._popt
