### 2021-03-02 30a8992
set minimum version of iminuit to 2.0
(cherry picked from commit 227fd21d688c06d18b745f52075e5ff19186e41b)
