+++ setup.py
@@ -28,7 +28,7 @@ setup(name='FitBenchmarking',
                         'configparser'],
       extras_require={'DFO': ['DFO-LS', 'dfogn'],
                       'SAS': ['sasmodels'],
-                      'minuit': ['iminuit<2'],
+                      'minuit': ['iminuit>=2'],
                       'bumps': ['bumps']},
       zip_safe=False,
       use_scm_version={'fallback_version': '0.1.0'},
