+++ Docker/BasicInstall/Dockerfile
@@ -1,5 +1,5 @@
 # Use ubuntu as a parent image
-FROM ubuntu:18.04
+FROM ubuntu:20.04
 
 # Install the packages we need to build fitbenchmarking
 RUN apt-get update && apt-get install -y \
