### 2021-03-03 #3080 @dhhagan
Hi there @tyronerees - It appears that there are a few items missing from the [review items](https://joss.readthedocs.io/en/latest/review_criteria.html#review-items) list that JOSS requires/recommends. Specifically, I don't see any installation instructions, API documentation, example usage, community guidelines, or automated tests. It is possible that I am just not seeing them, in which case, please point me in their direction! It would also be helpful to list those resources directly in the Readme file which is the first thing users often look at.

Please let me know if you have any questions.
