### 2021-03-03 #3080 @tyronerees
Hi @dhhagan -- thanks for handling this submission.

**installation instructions** are at https://fitbenchmarking.readthedocs.io/en/latest/users/install_instructions/index.html
**API documentation, example usage**  Users interact with FitBenchmarking via a CLI, not API.  Basic details of how it is run are found at https://fitbenchmarking.readthedocs.io/en/latest/users/running.html (with the options being fully explored in the rest of the Users section https://fitbenchmarking.readthedocs.io/en/latest/users/index.html)
**community guidelines** are at https://fitbenchmarking.readthedocs.io/en/latest/contributors/guidelines.html
**automated tests** are run via GitHub Actions, https://github.com/fitbenchmarking/fitbenchmarking/actions, and the tests are documented at https://fitbenchmarking.readthedocs.io/en/latest/users/tests.html

I'll open an issue to point to these from the README.  Please let me know if you need any further information.
