### 2021-03-03 #3080 @danielskatz
@dhhagan - can you explain why you started a scope query?  I was expecting you to assign yourself as editor.
