### 2021-03-03 fitbenchmarking/fitbenchmarking#727 @tyronerees
Update README 
As pointed out in https://github.com/openjournals/joss-reviews/issues/3080#issuecomment-789697411, it's useful to have links to the installation instructions, example usage, community guidelines, automated tests, etc, in the README, so it's easy for new users to find this.  This could be unobtrusively done so as is makes sense across github/pypi/rtd if we followed the model in, e.g., https://github.com/scipy/scipy or https://github.com/numpy/numpy
