### 2021-03-03 2d6c79f
extend linting tests to cover all code
