+++ ci/linting_tests.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
-pylint -r y fitbenchmarking/parsing/* --exit-zero
-status=$?
-flake8 fitbenchmarking/parsing/*
-status=$(($status + $?))
+status=0
+
+for dir in cli controllers core cost_func jacobian parsing results_processing utils
+do 
+    pylint -r y fitbenchmarking/$dir/* --exit-zero
+    status=$(($status + $?))
+    flake8 fitbenchmarking/$dir/*
+    status=$(($status + $?))
+done
 
 if [[ $status != 0 ]]
 then
