### 2021-03-04 641c6ed
Remove rm line for original source as needed for symlink
