### 2021-03-05 #3080 @danielskatz
I'm going to remove the scope query - let me know if you think this is needed
