### 2021-03-08 06c2b6c
Fix issues from review
(cherry picked from commit 90c621f70d8f4124c6976788afd92fa3f3866c5a)
