### 2021-03-08 b1569cd
Remove minuit and dfo regression tests
