+++ fitbenchmarking/systests/linux_expected_results/all_parsers.txt
@@ -1,8 +1,8 @@
-                                bumps                    dfo                   gsl                    mantid                 minuit                 ralfit                     scipy              scipy_ls
-                               amoeba                  dfogn lmsder: scipy 2-point       BFGS: scipy 2-point                 minuit      gn: scipy 2-point               Nelder-Mead       lm-scipy-no-jac
-BENNETT5            1.608e-05 (1.001)           inf (inf)[3]  1.639e-05 (1.021)[1]         0.02056 (1281)[2]      2.477e-05 (1.543)          1.606e-05 (1)      1.653e-05 (1.029)[1]  1.904e-05 (1.186)[1]
-cubic, Start 1  1.119e-14 (2.899e+07)      1.788e-20 (46.31)         3.861e-22 (1)      370.8 (9.604e+23)[2]  2.352e-07 (6.091e+14)  6.723e-13 (1.741e+09)  6.267e-05 (1.623e+17)[1]         3.861e-22 (1)
-cubic, Start 2  1.146e-14 (2.969e+07)  1.312e-14 (3.397e+07)         3.861e-22 (1)  0.0005949 (1.541e+18)[2]  7.579e-06 (1.963e+16)  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
-cubic-fba       1.146e-14 (2.969e+07)  5.903e-13 (1.529e+09)         3.861e-22 (1)     3.306e-06 (8.563e+15)  7.579e-06 (1.963e+16)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
-EMU 73673           1.032e+05 (98.62)               1046 (1)       1053 (1.006)[1]           1553 (1.484)[2]               1046 (1)               1046 (1)              1055 (1.009)              1046 (1)
-Problem Def 1   6.784e+05 (5.225e+15)          1.299e-10 (1)   0.03966 (3.055e+08)        375 (2.888e+12)[2]       4.214e-07 (3245)     0.1119 (8.618e+08)      3.837e-09 (29.55)[1]   0.03966 (3.055e+08)
+                                bumps                   gsl                    mantid                 ralfit                     scipy              scipy_ls
+                               amoeba lmsder: scipy 2-point       BFGS: scipy 2-point      gn: scipy 2-point               Nelder-Mead       lm-scipy-no-jac
+BENNETT5            1.608e-05 (1.001)  1.639e-05 (1.021)[1]         0.02056 (1281)[2]          1.606e-05 (1)      1.653e-05 (1.029)[1]  1.904e-05 (1.186)[1]
+cubic, Start 1  1.119e-14 (2.899e+07)         3.861e-22 (1)      370.8 (9.604e+23)[2]  6.723e-13 (1.741e+09)  6.267e-05 (1.623e+17)[1]         3.861e-22 (1)
+cubic, Start 2  1.146e-14 (2.969e+07)         3.861e-22 (1)  0.0005949 (1.541e+18)[2]  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
+cubic-fba       1.146e-14 (2.969e+07)         3.861e-22 (1)     3.306e-06 (8.563e+15)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
+EMU 73673           1.032e+05 (98.62)       1053 (1.006)[1]           1553 (1.484)[2]               1046 (1)              1055 (1.009)              1046 (1)
+Problem Def 1   6.784e+05 (5.225e+15)   0.03966 (3.055e+08)        375 (2.888e+12)[2]     0.1119 (8.618e+08)      3.837e-09 (29.55)[1]   0.03966 (3.055e+08)
