+++ ci/prep_and_tag_release.sh
@@ -42,7 +42,7 @@ sed -i -e "s|\(fitbenchmarking/Build%20and%20Publish\)[^\?]*|\1/${tag}|" \
     -e "s|\(readthedocs.io/en\).*\(/users/install_instructions/index.html\)|\1/${tag}\2|" \
     -e "s|\(readthedocs.io/en\).*\(/users/index.html\)|\1/${tag}\2|" \
     -e "s|\(readthedocs.io/en\).*\(/contributors/guidelines.html\)|\1/${tag}\2|" \
-    -e "s|\(readthedocs.io/en\).*\(/tests.html\)|\1/${tag}\2|" \
+    -e "s|\(readthedocs.io/en\).*\(/users/tests.html\)|\1/${tag}\2|" \
     README.md
 
 git add README.md
