### 2021-03-09 79db262
Update fitbenchmarking/cost_func/poisson_cost_func.py
Co-authored-by: Tyrone Rees <tyrone.rees@stfc.ac.uk>
