### 2021-03-10 f604181
Issue #490: Direct user to options in poisson docs
