### 2021-03-10 742c0c7
changes to avoid errors from conflicting options
