+++ fitbenchmarking/cost_func/weighted_nlls_cost_func.py
@@ -37,6 +37,8 @@ class WeightedNLLSCostFunc(BaseNLLSCostFunc):
         #: Container cached residual evaluation
         self.cache_rx = {'params': None, 'value': None}
 
+        self.invalid_algorithm_types = [None]
+
     def eval_r(self, params, **kwargs):
         """
         Calculate residuals
