+++ fitbenchmarking/core/fitting_benchmarking.py
@@ -18,7 +18,7 @@ from fitbenchmarking.cost_func.cost_func_factory import create_cost_func
 from fitbenchmarking.parsing.parser_factory import parse_problem_file
 from fitbenchmarking.utils.exceptions import NoResultsError, \
     UnknownMinimizerError, UnsupportedMinimizerError, MissingSoftwareError, \
-    ControllerAttributeError
+    ControllerAttributeError, NoJacobianError, IncompatibleMinimizerError
 from fitbenchmarking.jacobian.jacobian_factory import create_jacobian
 from fitbenchmarking.utils import fitbm_result, misc, output_grabber
 
@@ -123,18 +123,6 @@ def loop_over_benchmark_problems(problem_group, options):
                     cost_func, options, grabbed_output)
             results.extend(problem_results)
             failed_problems.extend(problem_fails)
-
-    # If the results are an empty list then this means that all minimizers
-    # raise an exception and the tables will produce errors if they run.
-    if results == []:
-        message = "The user chosen options and/or problem setup resulted in " \
-                  "all minimizers and/or parsers raising an exception. " \
-                  "This is likely due to the way `algorithm_type` was set " \
-                  "in the options or the selected problem set requires " \
-                  "additional software to be installed. Please review your " \
-                  "options setup and/or problem set then re-run " \
-                  "FitBenmarking."
-        raise NoResultsError(message)
 
     return results, failed_problems, unselected_minimzers, \
         minimizer_dict, cost_func_description
@@ -287,6 +275,13 @@ def loop_over_minimizers(controller, minimizers, options, grabbed_output):
             minimizer_check = False
             LOGGER.warn(str(excp))
 
+        try: 
+            controller.cost_func.validate_algorithm_type(controller.algorithm_check,minimizer)
+        except IncompatibleMinimizerError as excp:
+            minimizer_failed.append(minimizer)
+            minimizer_check = False
+            LOGGER.warn(str(excp))
+
         if minimizer_check:
             ########################
             # Loops over Jacobians #
@@ -341,6 +336,7 @@ def loop_over_jacobians(controller, options, grabbed_output):
     jacobian_list = options.jac_method
     minimizer_name = minimizer
     results = []
+    chi_sq = []
     new_minimizer_list = []
     minimizer_check = has_jacobian and minimizer not in invalid_jacobians
     try:
@@ -354,68 +350,76 @@ def loop_over_jacobians(controller, options, grabbed_output):
                         + (num_method if jac_method != "analytic" else '')
                 # Creates Jacobian class
                 jacobian_cls = create_jacobian(jac_method)
-                jacobian = jacobian_cls(cost_func)
-                jacobian.method = num_method
-
-                controller.jacobian = jacobian
                 try:
-                    with grabbed_output:
-                        # Calls timeit repeat with repeat = num_runs and
-                        # number = 1
-                        runtime_list = \
-                            timeit.Timer(setup=controller.prepare,
-                                         stmt=controller.fit).repeat(
-                                num_runs, 1)
-                        runtime = sum(runtime_list) / num_runs
-                        controller.cleanup()
-                        controller.check_attributes()
-                    ratio = np.max(runtime_list) / np.min(runtime_list)
-                    tol = 4
-                    if ratio > tol:
-                        warnings.warn(
-                            'The ratio of the max time to the min is {0}'
-                            ' which is  larger than the tolerance of {1},'
-                            ' which may indicate that caching has occurred'
-                            ' in the timing results'.format(ratio, tol))
-                    chi_sq = controller.eval_chisq(
-                        params=controller.final_params,
-                        x=controller.data_x,
-                        y=controller.data_y,
-                        e=controller.data_e)
-
[…]

[…]
