### 2021-03-12 694b4cf
remove levmar from default controllers
The levmar pypi package actually needs libblas and liblapack to be
available on the system, so can't be installed 'out of the box'
using pip.
