### 2021-03-12 bb38a88
Update fitbenchmarking/cost_func/poisson_cost_func.py
Co-authored-by: Tyrone Rees <tyrone.rees@stfc.ac.uk>
