### 2021-03-12 1b01b81
remove from default ci install, and document this
