+++ .github/workflows/main.yml
@@ -27,7 +27,7 @@ jobs:
     - name: Installing
       run: |
         sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install -r requirements.txt
-        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,minuit,SAS,numdifftools,levmar]
+        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,minuit,SAS,numdifftools]
     - name: 'Running Tests'
       run: |
         ci/system_tests_default.sh
@@ -43,7 +43,7 @@ jobs:
     - name: Installing
       run: |
         sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install -r requirements.txt
-        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,minuit,SAS,numdifftools,levmar]
+        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,levmar,minuit,SAS,numdifftools]
         mkdir -p $MASTSIF
         mkdir -p $PYCUTEST_CACHE
     - name: 'Running Tests'
@@ -61,7 +61,7 @@ jobs:
     - name: Installing
       run: |
         sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install -r requirements.txt
-        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,minuit,SAS,numdifftools,levmar]
+        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,minuit,SAS,numdifftools]
     - name: 'Running Tests'
       run: |
         ci/unit_tests_default.sh
@@ -77,7 +77,7 @@ jobs:
     - name: Installing
       run: |
         sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install -r requirements.txt
-        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,minuit,SAS,numdifftools,levmar]
+        sudo -HE --preserve-env=PATH $VIRTUAL_ENV/bin/pip install .[bumps,DFO,levmar,minuit,SAS,numdifftools]
         mkdir -p $MASTSIF
         mkdir -p $PYCUTEST_CACHE
     - name: 'Running Tests'
@@ -103,7 +103,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip==20.0.2
         pip install -r requirements.txt
-        pip install .[bumps,DFO,minuit,SAS,numdifftools,levmar]
+        pip install .[bumps,DFO,minuit,SAS,numdifftools]
         sudo apt-get install libgsl-dev
         pip install pygsl
         pip install pycutest
