### 2021-03-12 f8115d8
move test_levmar from default tests
