### 2021-03-12 a25528b
linting fixes (for the whole file)
