+++ fitbenchmarking/cost_func/nlls_base_cost_func.py
@@ -4,7 +4,7 @@ Implements the base non-linear least squares cost function
 from numpy import dot
 
 from fitbenchmarking.cost_func.base_cost_func import CostFunc
-from fitbenchmarking.utils.exceptions import CostFuncError, IncompatibleMinimizerError
+from fitbenchmarking.utils.exceptions import CostFuncError
 from abc import ABCMeta, abstractmethod
 
 
@@ -37,9 +37,7 @@ class BaseNLLSCostFunc(CostFunc):
         #: Container cached residual evaluation
         self.cache_rx = {'params': None, 'value': None}
 
-        # Used to check whether the algorithm type of the 
-        # selected minimizer is incompatible with the cost function
-        self.invalid_algorithm_types = [None]
+        self.invalid_algorithm_types = []
 
     @abstractmethod
     def eval_r(self, params, **kwargs):
@@ -70,25 +68,3 @@ class BaseNLLSCostFunc(CostFunc):
         self.cache_cost_x['params'] = params
         self.cache_cost_x['value'] = dot(r, r)
         return self.cache_cost_x['value']
-
-    def validate_algorithm_type(self,algorithm_check,minimizer):
-        """
-        Helper function which checks that the algorithm type of the
-        selected minimizer from the options (options.minimizer)
-        is incompatible with the selected cost function
-
-        :param algorithm_check: dictionary object containing algorithm
-        types and minimizers for selected software
-        :type algorithm_check: dict
-        :param minimizer: string of minimizers selected from the options
-        :type minimizer: str
-        """
-
-        for alg_type in list(algorithm_check.keys()):
-            if minimizer in algorithm_check[alg_type]:
-                incompatible = alg_type in self.invalid_algorithm_types
-                if incompatible:
-                    message = 'The algorithm type of the selected ' \
-                            'minimizer, {}, is not compatible with ' \
-                            'the selected cost function'.format(minimizer)
-                    raise IncompatibleMinimizerError(message)
