+++ fitbenchmarking/cost_func/nlls_cost_func.py
@@ -38,8 +38,6 @@ class NLLSCostFunc(BaseNLLSCostFunc):
         #: Container cached residual evaluation
         self.cache_rx = {'params': None, 'value': None}
 
-        self.invalid_algorithm_types = [None]
-
     def eval_r(self, params, **kwargs):
         """
         Calculate residuals

[…]
