+++ .github/workflows/main.yml
@@ -125,7 +125,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        pip install .[bumps,DFO,minuit,SAS,numdifftools,levmar]
+        pip install .[bumps,DFO,minuit,SAS,numdifftools]
     - name: 'Running Tests'
       run: |
         bash ci/unit_tests_default.sh
