+++ examples/all_softwares.ini
@@ -2,7 +2,7 @@
 software: bumps
           dfo
           gsl
-	  levmar
+          levmar
           mantid
           minuit
           ralfit
