### 2021-03-16 3cd63fe
Issue #490: Add the poisson cost function. (#713)
* Issue #490: Add the poisson cost function.

* Issue #490: First draft of simulated poisson

* Add tests and example poisson results

* Fix typo

* Fix analytic tests for poisson cost function

* Fix safe_log test

* Fix safe log failing when passed integer values

* Fixing issues spotted by self review

* Update fitbenchmarking/cost_func/poisson_cost_func.py

Co-authored-by: Tyrone Rees <tyrone.rees@stfc.ac.uk>

* Issue #490: Direct user to options in poisson docs

* Update fitbenchmarking/cost_func/poisson_cost_func.py

Co-authored-by: Tyron
[…]
