+++ fitbenchmarking/cost_func/tests/test_cost_func.py
@@ -6,13 +6,15 @@ from collections import OrderedDict
 from unittest import TestCase
 
 import numpy as np
-
-from fitbenchmarking.cost_func.weighted_nlls_cost_func import \
-    WeightedNLLSCostFunc
-from fitbenchmarking.cost_func.nlls_cost_func import NLLSCostFunc
-from fitbenchmarking.cost_func.hellinger_nlls_cost_func import HellingerNLLSCostFunc
 from fitbenchmarking.cost_func.base_cost_func import CostFunc
 from fitbenchmarking.cost_func.cost_func_factory import create_cost_func
+from fitbenchmarking.cost_func.hellinger_nlls_cost_func import \
+    HellingerNLLSCostFunc
+from fitbenchmarking.cost_func.nlls_cost_func import NLLSCostFunc
+from fitbenchmarking.cost_func.poisson_cost_func import (PoissonCostFunc,
+                                                         _safe_a_log_b)
+from fitbenchmarking.cost_func.weighted_nlls_cost_func import \
+    WeightedNLLSCostFunc
 from fitbenchmarking.parsing.fitting_problem import FittingProblem
 from fitbenchmarking.utils import exceptions
 from fitbenchmarking.utils.options import Options
@@ -162,6 +164,63 @@ class TestHellingerNLLSCostFunc(TestCase):
         self.assertEqual(eval_result, 0.4194038580206052)
 
 
+class TestPoissonCostFunc(TestCase):
+    """
+    Class to test the PoissonCostFunc class
+    """
+
+    def setUp(self):
+        """
+        Setting up poisson cost function tests
+        """
+        self.options = Options()
+        fitting_problem = FittingProblem(self.options)
+        self.cost_function = PoissonCostFunc(fitting_problem)
+        fitting_problem.function = lambda x, p1: x + p1
+        self.x_val = np.array([1, 8, 11])  # 6, 13, 16
+        self.y_val = np.array([6, 10, 20])
+
+    def test_eval_cost_raise_error(self):
+        """
+        Test that eval_cost raises an error if inputs are bad.
+        """
+        with self.assertRaises(exceptions.CostFuncError):
+            _ = self.cost_function.eval_cost(
+                params=[5],
+                x=[2],
+                y=[1, 3, 5])
+
+    def test_eval_cost_correct(self):
+        """
+        Test that the eval cost function returns the correct value
+        """
+        eval_result = self.cost_function.eval_cost(params=[5],
+                                                   x=self.x_val,
+                                                   y=self.y_val)
+
+        # 6*(log(6) - log(6))
+        # + 10*(log(10) - log(13))
+        # + 20*(log(20) - log(16))
+        # - (6 - 6) - (10 - 13) - (20 - 16)
+        # == 30*log(5) - 10*log(13) - 30*log(2) - 1
+        self.assertEqual(eval_result, 0.8392283816092849)
+
+    def test_safe_a_log_b(self):
+        """
+        Test the safe_a_log_b function.
+        """
+        a = np.array([1, 2, 3, 0, 5])
+        b = np.array([1, 2, 3, 4, 5])
+        res = _safe_a_log_b(a, b)
+        self.assertTrue(np.isclose(res,
+                                   np.array([0.0,
+                                             2*np.log(2),
+                                             3*np.log(3),
+                                             0.0,
+                                             5*np.log(5)])
+                                   ).all())
+
+
 class FactoryTests(TestCase):
     """
     Tests for the cost function factory
@@ -173,7 +232,7 @@ class FactoryTests(TestCase):
         """
         self.options = Options()
 
-        valid = ['weighted_nlls', 'nlls', 'hellinger_nlls']
[…]
