+++ examples/benchmark_problems/Poisson/simulated_bumps.txt
@@ -1,0 +1,6 @@
+# FitBenchmark Problem
+software = 'Mantid'
+name = 'Simulated poisson (bumps)'
+description = 'A simulated dataset for testing poisson costfunctions, based on a simple simulation from bumps.'
+input_file = 'simulated_bumps.txt'
+function = 'name=UserFunction,Formula=scale*exp(-0.5*(x-centre)^2/width^2)+background,scale=1.0,centre=2.0,width=2.0,background=0.0'
