### 2021-03-16 #3080 @dhhagan
:wave:  @ziotom78, @diehlpk, @AmirHoseinHadian, & @johnsamualwrites, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
