### 2021-03-16 4776fce
680 Behaviour when incompatible options selected (#735)
Co-authored-by: Andrew Lister <48282025+AndrewLister-STFC@users.noreply.github.com>
