### 2021-03-17 6e6b575
Fix exit check
