### 2021-03-19 d7437cf
Update docs/source/extending/controllers.rst
Co-authored-by: Andrew Lister <48282025+AndrewLister-STFC@users.noreply.github.com>
