### 2021-03-19 043119c
Update docs/source/users/install_instructions/fitbenchmarking.rst
