### 2021-03-19 424e758
Update fitbenchmarking/controllers/levmar_controller.py
