### 2021-03-22 1e144ce
add bounds support to scipy software (draft)
