### 2021-03-22 72ff52e
Deal with incompatible bounded problems & minimizers (draft)
