+++ fitbenchmarking/core/fitting_benchmarking.py
@@ -283,6 +283,30 @@ def loop_over_minimizers(controller, minimizers, options, grabbed_output):
             minimizer_check = False
             LOGGER.warning(str(excp))
 
+        if controller.problem.value_ranges is not None:
+            try:
+                controller.check_minimizer_bounds()
+            except IncompatibleMinimizerError as excp:
+                minimizer_failed.append(minimizer)
+                minimizer_check = False
+                controller.flag = 4
+                LOGGER.warning(str(excp))
+                #dummy_results = [{'options': options,
+                #            'cost_func': controller.cost_func,
+                #            'jac': None,
+                #            'chi_sq': np.inf,
+                #            'runtime': np.inf,
+                #            'minimizer': minimizer,
+                #            'initial_params': controller.initial_params,
+                #            'params': controller.final_params,
+                #            'error_flag': controller.flag,
+                #            'name': problem.name}]
+                #for result in dummy_results:
+                #    individual_result = fitbm_result.FittingResult(
+                #        **result)
+                #    results_problem.append(individual_result)
+                #new_minimizer_list.append(minimizer)
+              
         if minimizer_check:
             ########################
             # Loops over Jacobians #
