+++ fitbenchmarking/results_processing/base_table.py
@@ -284,9 +284,10 @@ class Table:
         support_page_link = links[name]
         i = 0
         for l, v in zip(support_page_link, value.array):
-            tmp_link = os.path.relpath(path=l,
-                                       start=self.group_dir)
-            value.array[i] = '<a href="{0}">{1}</a>'.format(tmp_link, v)
+            if l != '':
+                tmp_link = os.path.relpath(path=l,
+                                           start=self.group_dir)
+                value.array[i] = '<a href="{0}">{1}</a>'.format(tmp_link, v)
             i += 1
         return value
 
