+++ examples/options_do_not_work.ini
@@ -1,0 +1,23 @@
+[FITTING]
+
+software: scipy
+          bumps
+          scipy_supports_bounds
+
+num_runs: 1
+
+[MINIMIZERS]
+
+scipy_supports_bounds: Powell
+
+scipy: Nelder-Mead
+
+bumps: amoeba
+
+[PLOTTING]
+
+make_plots: yes
+
+[LOGGING]
+
+external_output: debug
