+++ examples/options_work.ini
@@ -1,0 +1,20 @@
+[FITTING]
+
+software: scipy
+		  bumps
+
+num_runs: 1
+
+[MINIMIZERS]
+
+bumps: amoeba
+
+scipy: Nelder-Mead
+
+[PLOTTING]
+
+make_plots: yes
+
+[LOGGING]
+
+external_output: debug
