### 2021-03-22 01b40df
Fix handling of incompatible problems/minimizers
