+++ fitbenchmarking/utils/options.py
@@ -32,15 +32,14 @@ class Options(object):
                     'SteepestDescent', 'Trust Region'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
-         'scipy': ['Nelder-Mead', 'CG', 'BFGS', 'Newton-CG'],
-         'scipy_supports_bounds' : ['Powell','L-BFGS-B','TNC','SLSQP'],
-         'scipy_ls': ['lm-scipy-no-jac', 'lm-scipy'],
-         'scipy_ls_supports_bounds' : ['trf', 'dogbox']}
+         'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
+                   'Newton-CG', 'L-BFGS-B', 'TNC', 'SLSQP'],
+         'scipy_ls': ['lm-scipy-no-jac', 'lm-scipy', 'trf',
+                      'dogbox']}
     VALID_FITTING = \
         {'algorithm_type': ['all', 'ls', 'deriv_free', 'general'],
          'software': ['bumps', 'dfo', 'gsl', 'mantid', 'minuit',
-                      'ralfit', 'scipy', 'scipy_supports_bounds',
-                      'scipy_ls', 'scipy_ls_supports_bounds'],
+                      'ralfit', 'scipy', 'scipy_ls'],
          'jac_method': ['scipy', 'analytic', 'numdifftools'],
          'cost_func_type': ['nlls', 'weighted_nlls', 'hellinger_nlls',
                             'poisson']}
@@ -80,10 +79,10 @@ class Options(object):
                     'SteepestDescent', 'Trust Region'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
-         'scipy': ['Nelder-Mead', 'CG', 'BFGS', 'Newton-CG'],
-         'scipy_supports_bounds' : ['Powell','L-BFGS-B','TNC','SLSQP'],
-         'scipy_ls': ['lm-scipy-no-jac', 'lm-scipy'],
-         'scipy_ls_supports_bounds' : ['trf', 'dogbox']}
+         'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
+                   'Newton-CG', 'L-BFGS-B', 'TNC', 'SLSQP'],
+         'scipy_ls': ['lm-scipy-no-jac', 'lm-scipy', 'trf',
+                      'dogbox']}
     DEFAULT_FITTING = \
         {'num_runs': 5,
          'algorithm_type': 'all',
