### 2021-03-24 #3127 @johnsamuelwrites
@tyronerees Is it possible to add a section 'State of the field' in the article?

"State of the field: How FitBenchmarking software compares to other commonly-used packages?"
