### 2021-03-24 9860cf7
document that adding controller means adding reg test
