+++ docs/source/extending/controllers.rst
@@ -55,7 +55,10 @@ repository so that other can use this package.  To do this need to follow our
    Unless the new controller is more complicated than the currently available
    controllers, this can be done by following the example of the others.
 
-6. If `pip` installable add to ``install_requires`` in ``setup.py`` and
+6. If the software is deterministic, add the software to the regression tests in
+   ``fitbenchmarking/systests/test_regression.py``. 
+   
+7. If `pip` installable add to ``install_requires`` in ``setup.py`` and
    add to the installation step in ``.github/workflows/release.yml``.
    If not, document the installation procedure in :ref:`external-instructions`
    and update the ``FullInstall`` Docker Container -- the main developers will
