### 2021-03-24 4c324be
add levmar as a regression test
