+++ fitbenchmarking/systests/linux_expected_results/all_parsers.txt
@@ -1,8 +1,8 @@
-                                bumps                   gsl                    mantid                 ralfit                     scipy              scipy_ls
-                               amoeba lmsder: scipy 2-point       BFGS: scipy 2-point      gn: scipy 2-point               Nelder-Mead       lm-scipy-no-jac
-BENNETT5            1.608e-05 (1.001)  1.639e-05 (1.021)[1]         0.02056 (1281)[2]          1.606e-05 (1)      1.653e-05 (1.029)[1]  1.904e-05 (1.186)[1]
-cubic, Start 1  1.119e-14 (2.899e+07)         3.861e-22 (1)      370.8 (9.604e+23)[2]  6.723e-13 (1.741e+09)  6.267e-05 (1.623e+17)[1]         3.861e-22 (1)
-cubic, Start 2  1.146e-14 (2.969e+07)         3.861e-22 (1)  0.0005949 (1.541e+18)[2]  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
-cubic-fba       1.146e-14 (2.969e+07)         3.861e-22 (1)     3.306e-06 (8.563e+15)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
-EMU 73673           1.032e+05 (98.62)       1053 (1.006)[1]           1553 (1.484)[2]               1046 (1)              1055 (1.009)              1046 (1)
-Problem Def 1   6.784e+05 (1.768e+14)   0.03966 (1.034e+07)        375 (9.775e+10)[2]     0.1119 (2.917e+07)          3.837e-09 (1)[1]   0.03966 (1.034e+07)
+                                bumps                   gsl                levmar                    mantid                 ralfit                     scipy              scipy_ls
+                               amoeba lmsder: scipy 2-point levmar: scipy 2-point       BFGS: scipy 2-point      gn: scipy 2-point               Nelder-Mead       lm-scipy-no-jac
+BENNETT5            1.608e-05 (1.001)  1.639e-05 (1.021)[1]        0.02183 (1359)         0.02056 (1281)[2]          1.606e-05 (1)      1.653e-05 (1.029)[1]  1.905e-05 (1.186)[1]
+cubic, Start 1  1.119e-14 (2.899e+07)         3.861e-22 (1)     148.9 (3.857e+23)      370.8 (9.604e+23)[2]  6.723e-13 (1.741e+09)  6.267e-05 (1.623e+17)[1]         3.861e-22 (1)
+cubic, Start 2  1.146e-14 (2.969e+07)         3.861e-22 (1)   0.01778 (4.604e+19)  0.0005949 (1.541e+18)[2]  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
+cubic-fba       1.146e-14 (2.969e+07)         3.861e-22 (1)   0.01778 (4.604e+19)     3.306e-06 (8.563e+15)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
+EMU 73673           1.032e+05 (98.62)       1053 (1.007)[1]     2.444e+04 (23.36)           1552 (1.484)[2]               1046 (1)              1055 (1.009)              1046 (1)
+Problem Def 1   6.784e+05 (1.768e+14)   0.04153 (1.083e+07)     357.5 (9.318e+10)        375 (9.775e+10)[2]     0.1119 (2.917e+07)          3.837e-09 (1)[1]   0.03966 (1.034e+07)
