### 2021-03-24 fdb0044
use empty list, rather than `None`
