### 2021-03-24 e3fffdf
apply further changes suggested in review
