+++ fitbenchmarking/controllers/tests/test_controllers.py
@@ -6,14 +6,6 @@ from unittest import TestCase
 import pytest
 from pytest import test_type as TEST_TYPE
 
-from fitbenchmarking.cost_func.weighted_nlls_cost_func import \
-    WeightedNLLSCostFunc
-from fitbenchmarking.parsing.parser_factory import parse_problem_file
-from fitbenchmarking.utils import exceptions
-from fitbenchmarking.utils.options import Options
-from fitbenchmarking.jacobian.scipy_jacobian import Scipy
-
-from fitbenchmarking import mock_problems
 from fitbenchmarking.controllers.base_controller import Controller
 from fitbenchmarking.controllers.bumps_controller import BumpsController
 from fitbenchmarking.controllers.controller_factory import ControllerFactory
@@ -27,6 +19,15 @@ if TEST_TYPE != "default":
     from fitbenchmarking.controllers.levmar_controller import LevmarController
     from fitbenchmarking.controllers.mantid_controller import MantidController
     from fitbenchmarking.controllers.ralfit_controller import RALFitController
+
+from fitbenchmarking.cost_func.weighted_nlls_cost_func import \
+    WeightedNLLSCostFunc
+from fitbenchmarking.parsing.parser_factory import parse_problem_file
+from fitbenchmarking.utils import exceptions
+from fitbenchmarking.utils.options import Options
+from fitbenchmarking.jacobian.scipy_jacobian import Scipy
+
+from fitbenchmarking import mock_problems
 
 
 def make_cost_func(file_name='cubic.dat'):
@@ -385,7 +386,6 @@ class ExternalControllerTests(TestCase):
                                          True,
                                          ["levmar-no-jac"])
 
-        print(controller._info)
         controller._info = (0, 1, 2, "Stop by small Dp", 4, 5, 6)
         self.shared_tests.check_converged(controller)
         controller._info = (0, 1, 2, "Stopped by small gradient J^T e",
