### 2021-03-24 d40ccac
Remove nan option from colour scale
