### 2021-03-24 5eb7f5a
change expected results for all_parsers.txt
