+++ fitbenchmarking/systests/linux_expected_results/all_parsers.txt
@@ -5,4 +5,4 @@ cubic, Start 1  1.119e-14 (2.899e+07)         3.861e-22 (1)      370.8 (9.604e+2
 cubic, Start 2  1.146e-14 (2.969e+07)         3.861e-22 (1)  0.0005949 (1.541e+18)[2]  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
 cubic-fba       1.146e-14 (2.969e+07)         3.861e-22 (1)     3.306e-06 (8.563e+15)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
 EMU 73673           1.032e+05 (98.62)       1053 (1.006)[1]           1553 (1.484)[2]               1046 (1)              1055 (1.009)              1046 (1)
-Problem Def 1   6.784e+05 (1.711e+07)           0.03966 (1)             375 (9455)[2]         0.1119 (2.821)              inf (inf)[4]          inf (inf)[4]
+Problem Def 1           6.784e+05 (1)          inf (inf)[4]             inf (inf)[4]            inf (inf)[4]              inf (inf)[4]          inf (inf)[4]
