### 2021-03-24 5f7a42e
386 add levmar (#722)
* minimally working hooks into levmar

* semi-working interface

* remove prints and update entry points

* fix linting

* add tests

* more linting

* add levmar to install process

* document levmar

* remove levmar from default controllers

The levmar pypi package actually needs libblas and liblapack to be
available on the system, so can't be installed 'out of the box'
using pip.

* remove from default ci install, and document this

* move test_levmar from default tests

* linting fixes (for the whole file)

* remove levmar from windows test build

* tab -> spaces

* U
[…]
