+++ docs/source/users/options/minimizer_option.rst
@@ -179,7 +179,31 @@ The Mantid minimizers are set as follows:
 
 .. warning::
    The external package Mantid must be installed to use these minimizers.
-	    
+
+Levmar (``levmar``)
+-------------------
+
+The `levmar <http://users.ics.forth.gr/~lourakis/levmar/>`_ package
+which implements the Levenberg-Marquardt method for nonlinear least-squares.
+We interface via the python interface `available on PyPI <https://pypi.org/project/levmar/>`_.
+
+* Levenberg-Marquardt with supplied Jacobian (:code:`levmar`)  - the Levenberg-Marquardt method
+* Levenberg-Marquardt with no Jacobian passed (:code:`levmar-no-jac`)  - as above, but using levmar's approximate Jacobian
+
+The `levmar` minimizers are set as follows:
+
+.. code-block:: rst
+
+   [MINIMIZERS]
+   levmar: levmar
+           levmar-no-jac
+
+
+.. warning::
+   The additional dependency levmar must be installed for this to be available;
+   See :ref:`extra_dependencies`. This package also requires the BLAS and LAPACK
+   libraries to be present on the system.
+	   
 Minuit (``minuit``)
 -------------------
 

[…]
