+++ fitbenchmarking/systests/linux_expected_results/all_parsers.txt
@@ -5,4 +5,4 @@ cubic, Start 1  1.119e-14 (2.899e+07)         3.861e-22 (1)     148.9 (3.857e+23
 cubic, Start 2  1.146e-14 (2.969e+07)         3.861e-22 (1)   0.01778 (4.604e+19)  0.0005949 (1.541e+18)[2]  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
 cubic-fba       1.146e-14 (2.969e+07)         3.861e-22 (1)   0.01778 (4.604e+19)     3.306e-06 (8.563e+15)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
 EMU 73673           1.032e+05 (98.62)       1053 (1.006)[1]     2.444e+04 (23.36)           1553 (1.484)[2]               1046 (1)              1055 (1.009)              1046 (1)
-Problem Def 1   6.784e+05 (1.768e+14)   0.03966 (1.034e+07)     357.5 (9.318e+10)        375 (9.775e+10)[2]     0.1119 (2.917e+07)          3.837e-09 (1)[1]   0.03966 (1.034e+07)
+Problem Def 1   6.784e+05 (1.768e+14)          inf (inf)[4]          inf (inf)[4]              inf (inf)[4]           inf (inf)[4]              inf (inf)[4]          inf (inf)[4]
