### 2021-03-25 c122d6d
improve extensibility of check for new software
