+++ fitbenchmarking/controllers/scipy_ls_controller.py
@@ -24,6 +24,7 @@ class ScipyLSController(Controller):
         """
         super(ScipyLSController, self).__init__(cost_func)
         self._param_names = self.problem.param_names
+        self.support_for_bounds = True
         self.no_bounds_minimizers = ['lm-scipy-no-jac', 'lm-scipy']
         self._popt = None
         self.algorithm_check = {
