+++ fitbenchmarking/controllers/bumps_controller.py
@@ -31,7 +31,7 @@ class BumpsController(Controller):
         super(BumpsController, self).__init__(cost_func)
 
         self._param_names = self.problem.param_names
-
+        self.support_for_bounds = True
         self._func_wrapper = None
         self._fit_problem = None
         self._bumps_result = None
