+++ docs/source/users/problem_definition_files/native.rst
@@ -89,7 +89,7 @@ fit_ranges
 parameter_ranges
   This specifies upper and lower bounds for parameters in the problem.
 
-  Similarly to fit_ranges, it takes the form where the first number
+  Similarly to `fit_ranges`, it takes the form where the first number
   is the minimum in the range and the second is the maximum.
 
   Currently in Fitbenchmarking, problems with parameter_ranges can only
