+++ docs/source/users/problem_definition_files/native.rst
@@ -92,5 +92,5 @@ parameter_ranges
   Similarly to `fit_ranges`, it takes the form where the first number
   is the minimum in the range and the second is the maximum.
 
-  Currently in Fitbenchmarking, problems with parameter_ranges can only
+  Currently in Fitbenchmarking, problems with `parameter_ranges` can only
   be handled by SciPy fitting software.
