+++ fitbenchmarking/cli/main.py
@@ -141,16 +141,18 @@ def run(problem_sets, options_file='', debug=False):
             benchmark(options=options,
                       data_dir=data_dir)
 
-        # If the results are an empty list then this means that all minimizers
-        # raise an exception and the tables will produce errors if they run
-        # for that problem set.
+        # If a result has error flag 4 then the result contains dummy values,
+        # if this is the case for all results then output should not be 
+        # produced as results tables won't show meaningful values.
         all_dummy_results_flag = True
         for result in results:
             for res in result:
                 if res.error_flag != 4:
                     all_dummy_results_flag = False
                     continue
-
+        # If the results are an empty list then this means that all minimizers
+        # raise an exception and the tables will produce errors if they run
+        # for that problem set.
         if results == [] or all_dummy_results_flag is True :
             message = "\nWARNING: \nThe user chosen options and/or problem " \
                       " setup resulted in all minimizers and/or parsers " \
