### 2021-03-26 e458f03
736 scipy support for bounds (#739)
* add bounds support to scipy software (draft)

* Deal with incompatible bounded problems & minimizers (draft)

* options files to test commented out code from previous commit

* Fix handling of incompatible problems/minimizers

* Corrections to controllers

* small changes

* Add parameter_ranges documentation

* add parameter bounds tests (using sasview example but could change to simpletests)

* Update regression tests expected results

* Remove nan option from colour scale

* change expected results for all_parsers.txt

* fix spacing

* add missing scipy
[…]
