+++ fitbenchmarking/controllers/scipy_ls_controller.py
@@ -4,6 +4,7 @@ In particular, for the scipy least_squares solver.
 """
 
 from scipy.optimize import least_squares
+import numpy as np
 
 from fitbenchmarking.controllers.base_controller import Controller
 
@@ -22,6 +23,9 @@ class ScipyLSController(Controller):
                 :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
         """
         super(ScipyLSController, self).__init__(cost_func)
+        self._param_names = self.problem.param_names
+        self.support_for_bounds = True
+        self.no_bounds_minimizers = ['lm-scipy-no-jac', 'lm-scipy']
         self._popt = None
         self.algorithm_check = {
             'all': ['lm-scipy-no-jac', 'lm-scipy', 'trf', 'dogbox'],
@@ -41,8 +45,26 @@ class ScipyLSController(Controller):
         """
         Setup problem ready to be run with SciPy LS
         """
+
         if self.minimizer == "lm-scipy":
             self.minimizer = "lm"
+
+        # If parameter ranges have been set in problem, then set up bounds
+        # option for scipy least_squares function. Here the bounds option
+        # must be a 2 tuple array like object, the first tuple containing
+        # the lower bounds for each parameter and the second containing all
+        # upper bounds.
+        value_ranges_lb = []
+        value_ranges_ub = []
+        for name in self._param_names:
+            if self.problem.value_ranges is not None \
+                    and name in self.problem.value_ranges:
+                value_ranges_lb.extend([self.problem.value_ranges[name][0]])
+                value_ranges_ub.extend([self.problem.value_ranges[name][1]])
+            else:
+                value_ranges_lb.extend([-np.inf])
+                value_ranges_ub.extend([np.inf])
+        self.value_ranges = (value_ranges_lb, value_ranges_ub)
 
     def fit(self):
         """
@@ -59,11 +81,18 @@ class ScipyLSController(Controller):
                                         x0=self.initial_params,
                                         method="lm",
                                         max_nfev=500)
+        elif self.minimizer == "lm":
+            self.result = least_squares(fun=self.cost_func.eval_r,
+                                        x0=self.initial_params,
+                                        method=self.minimizer,
+                                        jac=self.jacobian.eval,
+                                        max_nfev=500)
         else:
             self.result = least_squares(fun=self.cost_func.eval_r,
                                         x0=self.initial_params,
                                         method=self.minimizer,
                                         jac=self.jacobian.eval,
+                                        bounds=self.value_ranges,
                                         max_nfev=500)
 
         self._popt = self.result.x
