### 2021-03-29 4097762
Add support for bounds for minuit, ralfit and dfo
