+++ fitbenchmarking/controllers/dfo_controller.py
@@ -25,6 +25,8 @@ class DFOController(Controller):
         """
         super(DFOController, self).__init__(cost_func)
 
+        self.support_for_bounds = True
+        self._param_names = self.problem.param_names
         self._soln = None
         self._popt = None
         self._pinit = None
@@ -48,16 +50,48 @@ class DFOController(Controller):
         """
         self._pinit = np.asarray(self.initial_params)
 
+        value_ranges_lb = np.array([])
+        value_ranges_ub = np.array([])
+        bound_range = np.array([])
+        for name in self._param_names:
+            if self.problem.value_ranges is not None \
+                    and name in self.problem.value_ranges:
+                value_ranges_lb = np.append(value_ranges_lb, \
+                    self.problem.value_ranges[name][0])
+                value_ranges_ub = np.append(value_ranges_ub, \
+                    self.problem.value_ranges[name][1])
+                bound_range = np.append(bound_range, \
+                    self.problem.value_ranges[name][1] \
+                    - self.problem.value_ranges[name][0])
+            else:
+                value_ranges_lb = np.append(value_ranges_lb, -10e+20)
+                value_ranges_ub = np.append(value_ranges_ub, 10e+20)
+        self.value_ranges = (value_ranges_lb, value_ranges_ub)
+
+
+        # if bounds are set then gap between lower and upper bound must
+        # be at least 2*rhobeg so check that default rhobeg value
+        # satisfies this condition
+        self.rhobeg = 0.1*max(np.linalg.norm(self._pinit, np.inf),1)
+        if bound_range.size:
+            if min(bound_range) <= 2*self.rhobeg:
+                self.rhobeg = min(bound_range/2)
+
     def fit(self):
         """
         Run problem with DFO.
         """
         if self.minimizer == 'dfogn':
             self._soln = dfogn.solve(self.cost_func.eval_r,
-                                     self._pinit)
+                                     self._pinit,
+                                     rhobeg = self.rhobeg,
+                                     lower=self.value_ranges[0],
+                                     upper=self.value_ranges[1])
         elif self.minimizer == 'dfols':
             self._soln = dfols.solve(self.cost_func.eval_r,
-                                     self._pinit)
+                                     self._pinit,
+                                     rhobeg = self.rhobeg,
+                                     bounds=(self.value_ranges))
 
         self._popt = self._soln.x
         self._status = self._soln.flag
