### 2021-03-30 9be175e
bounds support for mantid problems
