+++ fitbenchmarking/controllers/minuit_controller.py
@@ -35,7 +35,6 @@ class MinuitController(Controller):
         super(MinuitController, self).__init__(cost_func)
 
         self.support_for_bounds = True
-        self._param_names = self.problem.param_names
         self._popt = None
         self._initial_step = None
         self._minuit_problem = None
