+++ fitbenchmarking/controllers/dfo_controller.py
@@ -26,7 +26,6 @@ class DFOController(Controller):
         super(DFOController, self).__init__(cost_func)
 
         self.support_for_bounds = True
-        self._param_names = self.problem.param_names
         self._soln = None
         self._popt = None
         self._pinit = None
