### 2021-03-31 8fa6e9d
pull more of bounds info into the problem class
