### 2021-03-31 35a8dc3
add unit test for set_value_ranges
