+++ fitbenchmarking/parsing/tests/test_fitting_problem.py
@@ -96,6 +96,17 @@ class TestFittingProblem(TestCase):
         function_def = fitting_problem.get_function_params(params=params)
         self.assertEqual(function_def, expected_function_def)
 
+    def test_set_value_ranges(self):
+        """
+        Tests that value_ranges are formatted correctly
+        """
+        fitting_problem = FittingProblem(self.options)
+        fitting_problem._param_names = ['param1','param2','param3']
+        value_ranges_prob_def = {'param1':(0,5),'param2':(5,10)}
+        expected_value_ranges = [(0,5),(5,10),(-np.inf,np.inf)]
+        fitting_problem.set_value_ranges(value_ranges_prob_def)
+        self.assertEqual(fitting_problem.value_ranges, expected_value_ranges)
+
     def test_correct_data_single_fit(self):
         """
         Tests that correct data gives the expected result
