+++ fitbenchmarking/parsing/tests/test_fitting_problem.py
@@ -101,7 +101,7 @@ class TestFittingProblem(TestCase):
         Tests that value_ranges are formatted correctly
         """
         fitting_problem = FittingProblem(self.options)
-        fitting_problem._param_names = ['param1','param2','param3']
+        fitting_problem.starting_values = [OrderedDict([('param1',0),('param2',0),('param3',0)])]
         value_ranges_prob_def = {'param1':(0,5),'param2':(5,10)}
         expected_value_ranges = [(0,5),(5,10),(-np.inf,np.inf)]
         fitting_problem.set_value_ranges(value_ranges_prob_def)
