+++ fitbenchmarking/parsing/fitting_problem.py
@@ -225,12 +225,12 @@ class FittingProblem:
         """
         if value_ranges is not None:
             self.value_ranges = []
-            for name in self._param_names:
+            for name in self.starting_values[0].keys():
                 param_name = name.lower()
                 if param_name in value_ranges:
                     self.value_ranges.append(
                         (value_ranges[param_name][0], value_ranges[param_name][1]))
-                elif value_ranges is not None:
+                else:
                     self.value_ranges.append((-np.inf, np.inf))
 
 
