+++ fitbenchmarking/controllers/levmar_controller.py
@@ -49,10 +49,10 @@ class LevmarController(Controller):
         """
 
         if self.value_ranges is not None:
-            lb,ub = zip(*self.value_ranges)
-            lb = [None if x==-np.inf else x for x in list(lb)]
-            ub = [None if x==np.inf else x for x in list(ub)]
-            self.param_ranges = list(zip(tuple(lb),tuple(ub)))
+            lb, ub = zip(*self.value_ranges)
+            lb = [None if x == -np.inf else x for x in list(lb)]
+            ub = [None if x == np.inf else x for x in list(ub)]
+            self.param_ranges = list(zip(tuple(lb), tuple(ub)))
 
     def _feval(self, p, x):
         """
