+++ fitbenchmarking/parsing/fitting_problem.py
@@ -229,7 +229,8 @@ class FittingProblem:
                 param_name = name.lower()
                 if param_name in value_ranges:
                     self.value_ranges.append(
-                        (value_ranges[param_name][0], value_ranges[param_name][1]))
+                        (value_ranges[param_name][0], 
+                            value_ranges[param_name][1]))
                 else:
                     self.value_ranges.append((-np.inf, np.inf))
 
