### 2021-04-04 #3127 @johnsamuelwrites
I covered all the items in the review checklist. Overall, fitbenchmarking is a well-documented software with commits from multiple contributors. The codebase contains necessary test cases and continuous integration workflows. The different modules are adequately structured and well-commented/well-documented. Thanks to the examples, new users can quickly start working and testing the software.

Further scope for improvement:
The documentation for the complete installation of the software with external dependencies for testing of all the functionalities needs some improvement. Currently, the users/testers may need to refer ".github/workflows/main.yml".

Also, please refer to my previous comment on "State of the field".
