### 2021-04-07 47e7828
Add Mantid CrystalField example
