+++ paper/paper.bib
@@ -82,3 +82,34 @@ keywords = {Benchmarking, CUTE, CUTEr, Modeling, CUTEst, Optimization}
   year = 2009
 }
 
+@misc{olympus,
+      title={Olympus: a benchmarking framework for noisy optimization and experiment planning}, 
+      author={Florian Häse and Matteo Aldeghi and Riley J. Hickman and Loïc M. Roch and Melodie Christensen and Elena Liles and Jason E. Hein and Alán Aspuru-Guzik},
+      year={2021},
+      eprint={2010.04153},
+      archivePrefix={arXiv},
+      primaryClass={stat.ML}
+}
+
+@article{paver,
+  title = {{{PAVER}} 2.0: An Open Source Environment for Automated Performance Analysis of Benchmarking Data},
+  shorttitle = {{{PAVER}} 2.0},
+  author = {Bussieck, Michael R. and Dirkse, Steven P. and Vigerske, Stefan},
+  year = {2014},
+  month = jul,
+  volume = {59},
+  pages = {259--275},
+  issn = {0925-5001, 1573-2916},
+  doi = {10.1007/s10898-013-0131-5},
+  abstract = {In this paper we describe PAVER 2.0, an environment (i.e. a process and a suite of tools supporting that process) for the automated performance analysis of benchmarking data. This new environment improves on its predecessor by addressing some of the shortcomings of the original PAVER [6] and extending its capabilities. The changes serve to further the original goals of PAVER (automation of the visualization and summarization of benchmarking data) while making the environment more accessible for the use of and modification by the entire community of potential users. In particular, we have targeted the end-users of optimization software, as they are best able to make the many subjective choices necessary to produce impactful results when benchmarking optimization software. We illustrate with some sample analyses conducted via PAVER 2.0.},
+  file = {C\:\\Users\\dgk98065\\Zotero\\storage\\HHCECWGY\\Bussieck et al. - 2014 - PAVER 2.0 an open source environment for automate.pdf},
+  journal = {Journal of Global Optimization},
+  language = {en},
+  number = {2-3}
+}
+
+@misc{benchopt,
+  title = {BenchOpt 1.1.0},
+  howpublished = {\url{https://benchopt.github.io},
+  note = {Accessed: 2021-07-04}
+}
