+++ paper/paper.md
@@ -54,7 +54,39 @@ Representatives of each of these communities have got together to build `FitBenc
 
 `FitBenchmarking` originally started as a tool to benchmark fitting algorithms in the data reduction package `Mantid` [@mantid], which is used to process neutron scattering and muon spectroscopy data. `FitBenchmarking` has since been significantly extended to take data and models from other real world applications and data analysis / modelling / treatment packages, such as `SasView` [@sasview] and `CUTEst` [@cutest]. It fits models to the data by using a range of data fitting and nonlinear optimization software packages, and present comparisons through a variety of different metrics. These include comparison tables and performance profile plots.
 
-`FitBenchmarking` compares how different fitting algorithms perform for the same data, model and initial guess. The best parameters for the model are found by solving a nonlinear least-squares problem, which can either be solved via a fitting package or an off the shelf minimizer. Figure \ref{fig:sample} displays a data set from `FitBenchmarking` where the crosses are the data points and the two curves are the fits found by two optimization algorithms implemented in `GSL` [@gsl]. From Figure \ref{fig:sample}, it is clear that the solution given by lmsder is better. As the volume of data increases, and we do more and more data analysis algorithmically, it is increasingly important that we have the best algorithm without needing to check it by eye. `FitBenchmarking` generates HTML output that makes it easy to compare minimizers on a given problem set.
+As far as we are aware, `FitBenchmarking` is the only package that is designed
+specifically to interface directly with data analysis packages to test different
+implemenations of fitting algorithms.  Below we highlight several packages available that
+are closest to `FitBenchmarking` in spirit.
+- `CUTEst` [@cutest] is written in Fortran, but interfaces
+to optimization packages in many languages, offering a way to benchmark any problems
+written in a custom SIF format.  However, only the hooks to run the same problem are
+provided, the user must provide their own data analysis.
+- Paver [@paver] is a package that is part of the COIN-OR initiative.
+Paver takes in data that has been generated by running a set of benchmark problems
+with a set of solvers as a csv file, and produces an analysis of these (e.g., as
+HTML or text tables, or via performance profiles).  In a way that's compliementary to
+CUTEst, Paver doesn't come with benchmark problems or link into solvers, but gives a
+method to analyse trace file generated by a user's own tests.
+- Olympus [@olympus] is a benchmarking framework designed for experiment planning,
+which provides analytic benchmark problems, experimental datasets, and emulated datasets.
+These benchmark test sets define optimization problems,
+on which different optimization methods
+(called planners in Olympus) can be tested.  A range of optimization methods are
+provided, but the focus is more on testing different algorithms, rather than comparing
+the implementations of algorithms.  However, Olympus is designed to add further
+problems (called emulators), or optimization methods, so could be used to benchmark
+data fitting problems.  
+- Benchopt [@benchopt] is a package for general optimization benchmarking, but which
+is primarily used to benchmark data fitting using a range of cost functions.
+A set of benchmarks on a limited number of datasets are publically available,
+and new benchmarks, comprising a data set, cost function, and a set of optimizers
+on which to run, can be added. At the end a plot of optimization progress vs time
+is produced.
+
+We designed `FitBenchmarking` to compliment and build on these packages, specifically
+to benchmark optimization methods applied to data fitting problems.
[…]
