+++ paper/paper.bib
@@ -110,6 +110,6 @@ keywords = {Benchmarking, CUTE, CUTEr, Modeling, CUTEst, Optimization}
 
 @misc{benchopt,
   title = {BenchOpt 1.1.0},
-  howpublished = {\url{https://benchopt.github.io},
+  howpublished = {\url{https://benchopt.github.io}},
   note = {Accessed: 2021-07-04}
 }
