### 2021-04-07 b57b50b
fix bullets
