+++ paper/paper.bib
@@ -111,5 +111,6 @@ keywords = {Benchmarking, CUTE, CUTEr, Modeling, CUTEst, Optimization}
 @misc{benchopt,
   title = {BenchOpt 1.1.0},
   howpublished = {\url{https://benchopt.github.io}},
-  note = {Accessed: 2021-07-04}
+  note = {Accessed: 2021-07-04},
+  year = {2021}
 }
