+++ paper/paper.md
@@ -61,24 +61,24 @@ are closest to `FitBenchmarking` in spirit.
 
 * `CUTEst` [@cutest] is written in Fortran, but interfaces
 to optimization packages in many languages, offering a way to benchmark any problems
-written in a custom SIF format.  However, only the hooks to run the same problem are
+written in a custom `SIF` format.  However, only the hooks to run the same problem are
 provided, the user must provide their own data analysis.
-* Paver [@paver] is a package that is part of the COIN-OR initiative.
-Paver takes in data that has been generated by running a set of benchmark problems
+* `Paver` [@paver] is a package that is part of the COIN-OR initiative.
+`Paver` takes in data that has been generated by running a set of benchmark problems
 with a set of solvers as a csv file, and produces an analysis of these (e.g., as
 HTML or text tables, or via performance profiles).  In a way that's compliementary to
-CUTEst, Paver doesn't come with benchmark problems or link into solvers, but gives a
+`CUTEst`, `Paver` doesn't come with benchmark problems or link into solvers, but gives a
 method to analyse trace file generated by a user's own tests.
-* Olympus [@olympus] is a benchmarking framework designed for experiment planning,
+* `Olympus` [@olympus] is a benchmarking framework designed for experiment planning,
 which provides analytic benchmark problems, experimental datasets, and emulated datasets.
 These benchmark test sets define optimization problems,
 on which different optimization methods
-(called planners in Olympus) can be tested.  A range of optimization methods are
+(called planners in `Olympus`) can be tested.  A range of optimization methods are
 provided, but the focus is more on testing different algorithms, rather than comparing
-the implementations of algorithms.  However, Olympus is designed to add further
+the implementations of algorithms.  However, `Olympus` is designed to add further
 problems (called emulators), or optimization methods, so could be used to benchmark
-data fitting problems.  
-* Benchopt [@benchopt] is a package for general optimization benchmarking, but which
+data fitting problems.
+* `Benchopt` [@benchopt] is a package for general optimization benchmarking, but which
 is primarily used to benchmark data fitting using a range of cost functions.
 A set of benchmarks on a limited number of datasets are publically available,
 and new benchmarks, comprising a data set, cost function, and a set of optimizers
