+++ docs/source/users/options/jacobian_option.rst
@@ -42,6 +42,25 @@ Default is ``2-point``
     [JACOBIAN]
     scipy: 2-point
 
+.. _solver-jac:
+
+Solver Default Jacobian (:code:`solver-jac`)
+--------------------------------------------
+
+This uses the approximation of the Jacobian that is used by default in the solver,
+and will vary between solvers.  If the solver requires the user to pass a Jacobian,
+a warning will be printed to the screen and the :ref:`scipy-jac`_ 2-point
+approximation will be used.  The only options is:
+
+* ``default`` - use the default derivative approxmation provided by the software.
+
+Default is ``default``
+
+.. code-block:: rst
+
+    [JACOBIAN]
+    solver-jac: default
+    
 .. _numdifftools-jac:
 
 Numdifftools (:code:`numdifftools`)
