### 2021-04-09 0914176
Update paper.md
* shorten the review so as not to better fit in with the short JOSS article format
* changed in places `method` with `algorithm` to be more consistent with how we use these words elsewhere in article, and similar with `individual scientific software` and `optimization software` packages
* without changes the meaning of the paper, in a few places made a few other changes to attempt to further increase the readability slightly
