+++ docs/source/users/options/jacobian_option.rst
@@ -49,7 +49,7 @@ Solver Default Jacobian (:code:`solver-jac`)
 
 This uses the approximation of the Jacobian that is used by default in the solver,
 and will vary between solvers.  If the solver requires the user to pass a Jacobian,
-a warning will be printed to the screen and the :ref:`scipy-jac`_ 2-point
+a warning will be printed to the screen and the `scipy-jac`_ 2-point
 approximation will be used.  The only options is:
 
 * ``default`` - use the default derivative approxmation provided by the software.
