### 2021-04-09 c987fad
handle not implemented jacobian error in output tables
