### 2021-04-09 6626030
add default jac for scipy controller
