### 2021-04-09 1f40eaa
add default jacobian to scipy minimizers
