### 2021-04-12 952fcbb
Remove indexing
Co-authored-by: Andrew Lister <48282025+AndrewLister-STFC@users.noreply.github.com>
