+++ fitbenchmarking/controllers/tests/test_controllers.py
@@ -683,10 +683,8 @@ class ExternalControllerBoundsTests(TestCase):
         controller.fit()
         controller.cleanup()
 
-        lower = [controller.value_ranges[i][0]
-                 for i in range(len(controller.value_ranges))]
-        upper = [controller.value_ranges[i][1]
-                 for i in range(len(controller.value_ranges))]
+        lower = [value_range[0] for value_range in controller.value_ranges]
+        upper = [value_range[1] for value_range in controller.value_ranges]
 
         # Convert None values to -inf/inf
         lower = [-np.inf if x is None else x for x in lower]
