### 2021-04-12 0a903b1
Only set value_ranges if not None
Co-authored-by: Andrew Lister <48282025+AndrewLister-STFC@users.noreply.github.com>
