### 2021-04-12 2be14ce
Remove more unnecessary casting to list/tuple
