### 2021-04-12 fa13953
745 Pull more of the bounds info into the problem class (#748)
