### 2021-04-12 7f5d47b
Add support for Mantid attributes
