### 2021-04-13 301729c
Add support for bounds to mantid controller
