+++ docs/source/users/problem_definition_files/native.rst
@@ -95,4 +95,4 @@ parameter_ranges
   is the minimum in the range and the second is the maximum.
 
   Currently in Fitbenchmarking, problems with `parameter_ranges` can
-  be handled by SciPy, Minuit, DFO, Levmar and RALFit fitting software.
+  be handled by SciPy, Minuit, DFO, Levmar, Mantid and RALFit fitting software.
