### 2021-04-14 2874171
Revert create_mantid_function edits (not required)
