### 2021-04-14 c605337
Get mantid function string properly
