### 2021-04-14 37f718c
"update the .rst files and the minuit_controller.py docstring to address written expression errors"
