### 2021-04-14 ffdb653
Optimize over all non-fixed parameters
