### 2021-04-15 f6161cc
Updated documentation .rst files and the minuit_controller.py docstring to address written expression errors (#760)
