### 2021-04-15 20bfacb
Optimize over all parameters (fixed and non-fixed)
