### 2021-04-15 c696b7d
Issue #494: Add support for ivp functions (first draft)
