+++ fitbenchmarking/cost_func/nlls_cost_func.py
@@ -52,8 +52,8 @@ class NLLSCostFunc(BaseNLLSCostFunc):
         y = kwargs.get("y", self.problem.data_y)
         if len(x) != len(y):
             raise CostFuncError('The length of the x and y are not the same, '
-                                'len(x)={} and len(y)= {}.'.format(len(x),
-                                                                   len(y)))
+                                'len(x)={} and len(y)={}.'.format(len(x),
+                                                                  len(y)))
         result = y - self.problem.eval_model(params=params, x=x)
         self.cache_rx['params'] = params
         self.cache_rx['value'] = result
