### 2021-04-15 d53a019
Change mantid function handling to use factory
