+++ paper/paper.md
@@ -49,29 +49,46 @@ Representatives of each of these communities have got together to build `FitBenc
 
 `FitBenchmarking` is easy to install via `pip` and our [documentation](https://fitbenchmarking.com/) guides users through the installation of some external packages we support. We provide several data sets from a range of applications and adding new data in these formats is as easy as dropping the data into a new folder. The data and fitting packages currently supported are shown in Figure \ref{fig:concept}. A key part of `FitBenchmarking` is the ease of which a user, with a basic knowledge of `Python`, can add new fitting software, data formats and different fitting comparison output metrics.
 
+# State of the field
+
+For general optimization, `CUTEst` [@cutest] (and predecessors) has been the
+standard tool to compare packages for some time, and can benchmark any problems
+written in a custom `SIF` format.  However, only the hooks to run the same problem are
+provided, the user must provide their own data analysis.  Tools such as
+`Paver` [@paver], a package that is part of the COIN-OR initiative, can be used
+alongside `CUTEst` (or other tools) for this purpose.
+The packages `Olympus` [@olympus] and `Benchopt` [@benchopt] have been developed
+as benchmarking and analysis frameworks for optimization problem.
+`Olympus` is designed for experiment planning and provides analytic benchmark problems,
+experimental datasets, and emulated datasets, but could be adapted to be applied to
+any optimization (or data-fitting) problem.
+`Benchopt`, on the other hand, is currently primarily used to benchmark data fitting
+using a range of cost functions.  `Benchopt` ships with a limited number of example data
+sets, but it is well documented how to write new benchmarks with custom data and
+objective functions.
+
 
 # Statement of need
 
 `FitBenchmarking` originally started as a tool to benchmark fitting algorithms in the data reduction package `Mantid` [@mantid], which is used to process neutron scattering and muon spectroscopy data. `FitBenchmarking` has since been significantly extended to take data and models from other real world applications and data analysis / modelling / treatment packages, such as `SasView` [@sasview] and `CUTEst` [@cutest]. It fits models to the data by using a range of data fitting and nonlinear optimization software packages, and present comparisons through a variety of different metrics. These include comparison tables and performance profile plots.
 
+While there is some overlap between `FitBenchmarking` and the rest of the field,
+what makes our software unique is:
+
+* It is designed to interface directly to the source of data, be that a scientific
+  software package or an academic data set.  Our `parser` class can be extended
+  to make it clear what a developer needs to do to get data into `FitBenchmarking`.
+* While being easy to extend using new software, or new data from currently supported
+  packages, `FitBechmarking` ships with open datasets that all can use for testing.
+* FitBenchmarking tests implementations of algorithms, not just algorithms.
+  A growing number of packages that can be used for data fitting are supported, and
+  it is straightforward to extend our `controller` class to add new software.
+
 As far as we are aware, `FitBenchmarking` is the only package that is designed
-specifically to interface directly with optimization  packages and individual scientific software packages to test different
-implementations of fitting algorithms.  Below we highlight packages available that
-are closest to `FitBenchmarking` in spirit.
+specifically to interface directly with optimization  packages and individual
+scientific software packages to test different implementations of fitting algorithms.
 
-* `CUTEst` [@cutest] offers a way to benchmark any problems
-written in a custom `SIF` format.  However, only the hooks to run the same problem are
-provided, the user must provide their own data analysis.
-* `Paver` [@paver] is a package that is part of the COIN-OR initiative.
-`Paver` takes in data that has been generated by running a set of benchmark problems
-with a set of solvers as a csv file, and produces an analysis of these.
-* `Olympus` [@olympus] is a benchmarking framework designed for experiment planning,
[…]
