+++ paper/paper.md
@@ -51,14 +51,15 @@ Representatives of each of these communities have got together to build `FitBenc
 
 # State of the field
 
-For general optimization, `CUTEst` [@cutest] (and predecessors) has been the
-standard tool to compare packages for some time, and can benchmark any problems
+Fitting data to models is a form of optimization, and 
+for general optimization, `CUTEst` [@cutest], and its predecessors, has been the
+standard tool to benchmark packages for some time. `CUTEst` can benchmark any problem
 written in a custom `SIF` format.  However, only the hooks to run the same problem are
 provided, the user must provide their own data analysis.  Tools such as
-`Paver` [@paver], a package that is part of the COIN-OR initiative, can be used
+`Paver` [@paver], with is part of the COIN-OR initiative, can be used
 alongside `CUTEst` (or other tools) for this purpose.
-The packages `Olympus` [@olympus] and `Benchopt` [@benchopt] have been developed
-as benchmarking and analysis frameworks for optimization problem.
+The packages `Olympus` [@olympus] and `Benchopt` [@benchopt] have been recently
+developed as benchmarking and analysis frameworks for optimization problems.
 `Olympus` is designed for experiment planning and provides analytic benchmark problems,
 experimental datasets, and emulated datasets, but could be adapted to be applied to
 any optimization (or data-fitting) problem.
@@ -83,6 +84,8 @@ what makes our software unique is:
 * FitBenchmarking tests implementations of algorithms, not just algorithms.
   A growing number of packages that can be used for data fitting are supported, and
   it is straightforward to extend our `controller` class to add new software.
+* `FitBenchmarking` performs its own data analysis and processing, and the output
+  generated can be customized if appropriate for new data sets and/or minimizers.  
 
 As far as we are aware, `FitBenchmarking` is the only package that is designed
 specifically to interface directly with optimization  packages and individual
