### 2021-04-16 5fcfb05
749 state of field (#751)
* add state of the field data

* add whitespace at end of bib file

* fix missing }

* fix bullets

* fixes based on preview

* Update paper.md

* shorten the review so as not to better fit in with the short JOSS article format
* changed in places `method` with `algorithm` to be more consistent with how we use these words elsewhere in article, and similar with `individual scientific software` and `optimization software` packages
* without changes the meaning of the paper, in a few places made a few other changes to attempt to further increase the readability slightly

[…]
