### 2021-04-16 #3127 @kyleniemeyer
Great, thanks @tyronerees! One final thing I noticed: for the Arnold et al. reference, you give a URL (that I'm not sure I think is appropriate, if that is a journal citation), but is there a DOI for that paper?

After making that change, can you archive the software repository (e.g., on Zenodo), and provide the DOI here? That is the last thing we need before accepting.
