+++ paper/paper.bib
@@ -11,7 +11,7 @@
 
   Owner                    = {trees},
   Timestamp                = {2017.01.17},
-  doi                      = {https://doi.org/10.1016/j.nima.2014.07.029}
+  doi                      = {10.1016/j.nima.2014.07.029}
 }
 
 @article{cutest,
