### 2021-04-19 ac60b7b
use mantid ref from ScienceDirect
