+++ paper/paper.bib
@@ -1,17 +1,15 @@
-@Article{mantid,
-  Title                    = {Mantid-Data analysis and visualization package for neutron scattering and $\mu$SR experiments},
-  Author                   = {Arnold, O. and Bilheux, J.C. and Borreguero, J.M. and Buts, A. and Campbell, S.I. and Chapon, L. and Doucet, M. and Draper, N. and Ferraz Leal, R. and Gigg, M.A. and et al.},
-  Journal                  = {Nuclear Instruments and Methods in Physics Research Section A},
-  Year                     = {2014},
-
-  Month                    = {November},
-  Note                     = {www.mantidproject.org},
-  Pages                    = {156-166},
-  Volume                   = {764},
-
-  Owner                    = {trees},
-  Timestamp                = {2017.01.17},
-  doi                      = {10.1016/j.nima.2014.07.029}
+@article{mantid,
+title = {Mantid---{D}ata analysis and visualization package for neutron scattering and $\mu$ {SR} experiments},
+journal = {Nuclear Instruments and Methods in Physics Research Section A: Accelerators, Spectrometers, Detectors and Associated Equipment},
+volume = {764},
+pages = {156-166},
+year = {2014},
+issn = {0168-9002},
+doi = {10.1016/j.nima.2014.07.029},
+url = {https://www.sciencedirect.com/science/article/pii/S0168900214008729},
+author = {O. Arnold and J.C. Bilheux and J.M. Borreguero and A. Buts and S.I. Campbell and L. Chapon and M. Doucet and N. Draper and R. {Ferraz Leal} and M.A. Gigg and V.E. Lynch and A. Markvardsen and D.J. Mikkelson and R.L. Mikkelson and R. Miller and K. Palmen and P. Parker and G. Passos and T.G. Perring and P.F. Peterson and S. Ren and M.A. Reuter and A.T. Savici and J.W. Taylor and R.J. Taylor and R. Tolchenov and W. Zhou and J. Zikovsky},
+keywords = {Data analysis, Data visualization, Computer interfaces},
+abstract = {The Mantid framework is a software solution developed for the analysis and visualization of neutron scattering and muon spin measurements. The framework is jointly developed by software engineers and scientists at the ISIS Neutron and Muon Facility and the Oak Ridge National Laboratory. The objectives, functionality and novel design aspects of Mantid are described.}
 }
 
 @article{cutest,
