### 2021-04-19 3cefc4f
Change test data set from sasview to mantid
