### 2021-04-19 22f6fb7
Remove bounds tests from default as mantid needed
