+++ fitbenchmarking/controllers/tests/test_controllers.py
@@ -391,8 +391,8 @@ class DefaultControllerTests(TestCase):
         controller._status = -1
         self.shared_tests.check_diverged(controller)
 
-
-class DefaultControllerBoundsTests(TestCase):
+@pytest.mark.skipif("TEST_TYPE == 'default'")
+class ControllerBoundsTests(TestCase):
 
     def setUp(self):
         """
@@ -469,6 +469,41 @@ class DefaultControllerBoundsTests(TestCase):
 
         self.check_bounds(controller)
 
+    def test_ralfit(self):
+        """
+        RALFitController: Test that parameter bounds are
+        respected for bounded problems
+        """
+        controller = RALFitController(self.cost_func)
+        controller.minimizer = 'gn'
+        controller.jacobian = self.jac
+
+        self.check_bounds(controller)
+
+    def test_levmar(self):
+        """
+        LevmarController: Test that parameter bounds are
+        respected for bounded problems
+        """
+
+        controller = LevmarController(self.cost_func)
+        controller.minimizer = 'levmar'
+        controller.jacobian = self.jac
+
+        controller.parameter_set = 0
+        controller.prepare()
+        controller.fit()
+        controller.cleanup()
+
+        lower = [value_range[0] for value_range in controller.value_ranges]
+        upper = [value_range[1] for value_range in controller.value_ranges]
+
+        # Convert None values to -inf/inf
+        lower = [-np.inf if x is None else x for x in lower]
+        upper = [np.inf if x is None else x for x in upper]
+
+        for count, value in enumerate(controller.final_params):
+            assert lower[count] <= value <= upper[count]
 
 @pytest.mark.skipif("TEST_TYPE == 'default'")
 class ExternalControllerTests(TestCase):
@@ -639,70 +674,6 @@ class ExternalControllerTests(TestCase):
             self.shared_tests.check_converged(controller)
             controller._status = 2
             self.shared_tests.check_diverged(controller)
-
-
-@pytest.mark.skipif("TEST_TYPE == 'default'")
-class ExternalControllerBoundsTests(TestCase):
-
-    def setUp(self):
-        """
-        Setup for bounded problem
-        """
-        self.cost_func = make_cost_func('cubic-fba-test-bounds.txt')
-        self.problem = self.cost_func.problem
-        self.jac = Scipy(self.cost_func)
-        self.jac.method = '2-point'
-
-    def check_bounds(self, controller):
-        """
-        Run bounded problem and check `final_params` respect
-        parameter bounds
-        """
-        controller.parameter_set = 0
-        controller.prepare()
-        controller.fit()
-        controller.cleanup()
-
-        for count, value in enumerate(controller.final_params):
-            assert controller.value_ranges[count][0] <= value \
-                <= controller.value_ranges[count][1]
-
-    def test_ralfit(self):
-        """
-        RALFitController: Test that parameter bounds are
-        respected for bounded problems
-        """
-        controller = RALFitController(self.cost_func)
-        controller.minimizer = 'gn'
-        controller.jacobian = self.jac
-
-        self.check_bounds(controller)
-
-    def test_levmar(self):
-        """
-        LevmarController: Test that parameter bounds are
-        respected for bounded problems
-        """
-
-        controller = LevmarController(self.cost_func)
-        controller.minimizer = 'levmar'
-        controller.jacobian = self.jac
-
-        controller.parameter_set = 0
-        controller.prepare()
-        controller.fit()
-        controller.cleanup()
-
-        lower = [value_range[0] for value_range in controller.value_ranges]
-        upper = [value_range[1] for value_range in controller.value_ranges]
-
-        # Convert None values to -inf/inf
[…]
