### 2021-04-20 c196f6c
add jacobian options tests
