+++ fitbenchmarking/utils/tests/test_options_jacobian.py
@@ -27,7 +27,8 @@ class JacobianOptionTesJacobians(unittest.TestCase):
         """
         expected = {'scipy': ['2-point'],
                     'analytic': ['cutest'],
-                    'numdifftools': ['central']}
+                    'numdifftools': ['central'],
+                    'solver': ['default']}
         actual = self.options.num_method
         self.assertEqual(expected, actual)
 
@@ -113,7 +114,8 @@ class UserJacobianOptionTests(unittest.TestCase):
         """
         set_option = {'scipy': ['cs'],
                       'analytic': ['cutest'],
-                      'numdifftools': ['central']}
+                      'numdifftools': ['central'],
+                      'solver': ['default']}
         config_str = \
             "[JACOBIAN]\nscipy: cs"
         self.shared_valid('num_method', set_option, config_str)
