### 2021-04-20 eacc321
Make multifit param handling more robust
