+++ fitbenchmarking/controllers/mantid_controller.py
@@ -202,14 +202,13 @@ class MantidController(Controller):
             self._mantid_results.OutputParameters.column(0),
             self._mantid_results.OutputParameters.column(1))}
 
-        final_params = [final_params_dict[key] for key in self._param_names]
-
-        num_params = len(self.initial_params)
-        if self._multi_fit:
-            self.final_params = [final_params[i * num_params:(i + 1) * num_params]
-                                 for i in range(self._multi_fit)]
+        if not self._multi_fit:
+            self.final_params = [final_params_dict[key]
+                                 for key in self._param_names]
         else:
-            self.final_params = final_params
+            self.final_params = [[final_params_dict[f'f{i}.{name}']
+                                  for name in self._param_names]
+                                 for i in range(self._multi_fit)]
 
     # Override if multi-fit
     # =====================
