### 2021-04-20 7d7d191
add solver jac tests
