+++ fitbenchmarking/jacobian/tests/test_jacobian.py
@@ -10,6 +10,7 @@ from fitbenchmarking.utils.options import Options
 from fitbenchmarking.utils import exceptions
 from fitbenchmarking.jacobian.base_jacobian import Jacobian
 from fitbenchmarking.jacobian.scipy_jacobian import Scipy
+from fitbenchmarking.jacobian.solver_jacobian import solver
 from fitbenchmarking.jacobian.numdifftools_jacobian import numdifftools
 from fitbenchmarking.jacobian.analytic_jacobian import Analytic
 from fitbenchmarking.jacobian.jacobian_factory import create_jacobian
@@ -95,6 +96,13 @@ class TestJacobianClass(TestCase):
         eval_result = jac.eval(params=self.params)
         self.assertTrue(np.isclose(self.actual, eval_result).all())
 
+    def test_solver_default(self):
+        """
+        Test that solver default jacobian does what it should
+        """
+        jac = solver(self.cost_func)
+        self.assertTrue(jac.use_solver_jac)
+    
     def test_numdifftools_eval(self):
         """
         Test whether numdifftools evaluation is correct
@@ -219,7 +227,6 @@ class TestCachedFuncValues(TestCase):
                                                           self.func,
                                                           params)
         assert expected_value == computed_value
-
 
 class TestDerivCostFunc(TestCase):
     """
