### 2021-04-20 f38211e
Update tests for new mantid implementation
