### 2021-04-21 c147b69
remove *-no-jac option from codebase
