+++ docs/source/users/problem_definition_files/native.rst
@@ -96,5 +96,4 @@ parameter_ranges
 
   Currently in Fitbenchmarking, problems with `parameter_ranges` can
   be handled by SciPy, Bumps, Minuit, DFO, Levmar and RALFit fitting
-  software. Please note that for SASView problems, Bumps does not
-  necessarily respect parameter bounds.
+  software.
