### 2021-04-21 83c9899
Update test_tables to make it more user friendly to update
