### 2021-04-22 fcd0520
Add flag for minimizers not respecting bounds
