+++ fitbenchmarking/controllers/tests/test_controllers.py
@@ -284,6 +284,28 @@ class BaseControllerTests(TestCase):
         with self.assertRaises(exceptions.IncompatibleMinimizerError):
             controller.check_minimizer_bounds(minimizer)
 
+    def test_bounds_respected_true(self):
+
+        controller = DummyController(self.cost_func)
+        controller.value_ranges = [(10,20),(20,30)]
+        controller.final_params = [15,30]
+        controller.flag = 0
+
+        controller.check_bounds_respected()
+
+        assert controller.flag == 0
+
+    def test_bounds_respected_false(self):
+
+        controller = DummyController(self.cost_func)
+        controller.value_ranges = [(10,20),(20,30)]
+        controller.final_params = [25,35]
+        controller.flag = 0
+
+        controller.check_bounds_respected()
+
+        assert controller.flag == 5
+
 
 class DefaultControllerTests(TestCase):
     """
