### 2021-04-23 1ae5175
Make a note of the CrystalField-DampedGN err
