### 2021-04-23 156f842
770 fix bumps parameter ordering (#771)
* Ensure results list parameters in correct order

* Update expected results for regression tests
