### 2021-04-23 6b35df4
770 fix bumps parameter ordering (#771)
* Ensure results list parameters in correct order

* Update expected results for regression tests

(cherry picked from commit 156f8428d5cb0e1bf4ad988daab83370c677ce2f)
