+++ fitbenchmarking/controllers/bumps_controller.py
@@ -82,6 +82,14 @@ class BumpsController(Controller):
         # acceptable by Bumps fitting.
         self._func_wrapper = func_wrapper
         self._fit_problem = FitProblem(func_wrapper)
+
+        # Determine the order of the parameters in `self.fit_problem` as this
+        # could differ from the ordering of parameters in `self._param_names`
+        param_order = []
+        for i in range(len(self._param_names)):
+            param_order.append(str(self._fit_problem._parameters[i]))
+        self.fit_order = param_order
+
         if self.minimizer == "lm-bumps":
             self.minimizer = "lm"
 
@@ -106,4 +114,14 @@ class BumpsController(Controller):
         else:
             self.flag = 2
 
-        self.final_params = self._bumps_result.x
+        # Set result variable where parameters are in the same
+        # order that are listed in `self._param_names`
+        result = []
+        if self.fit_order != self._param_names:
+            for name in self._param_names:
+                ind = self.fit_order.index(name)
+                result.append(self._bumps_result.x[ind])
+        else:
+            result = self._bumps_result.x
+
+        self.final_params = result
