+++ docs/source/users/problem_definition_files/native.rst
@@ -85,3 +85,4 @@ fit_ranges
 
   It takes the form shown in the example, where the first number
   is the minimum in the range and the second is the maximum.
+
