+++ fitbenchmarking/jacobian/solver_jacobian.py
@@ -3,6 +3,7 @@ Module which uses the solver's default jacobian
 """
 from fitbenchmarking.jacobian.base_jacobian import Jacobian
 
+
 # pylint: disable=useless-super-delegation
 class solver(Jacobian):
     """
@@ -12,4 +13,3 @@ class solver(Jacobian):
     def __init__(self, cost_func):
         super(solver, self).__init__(cost_func)
         self.use_solver_jac = True
-
