### 2021-04-25 48fcb64
remove flake8 warning and error in main.py
