### 2021-04-25 4316a75
fix flake8 W and E in test_main.py
