### 2021-04-25 0dbc94e
fix pylint warning in main.py
