### 2021-04-25 4fa6f06
fixed pylint W and E elsewhere in cli
