### 2021-04-26 5ec48d9
Issue #494: Add data assimilation for fitting side
