+++ fitbenchmarking/cost_func/nlls_cost_func.py
@@ -55,6 +55,10 @@ class NLLSCostFunc(BaseNLLSCostFunc):
                                 'len(x)={} and len(y)={}.'.format(len(x),
                                                                   len(y)))
         result = y - self.problem.eval_model(params=params, x=x)
+
+        # Flatten in case of a vector function
+        result = result.flatten()
+
         self.cache_rx['params'] = params
         self.cache_rx['value'] = result
         return result
