+++ fitbenchmarking/parsing/fitbenchmark_parser.py
@@ -4,13 +4,12 @@ This file implements a parser for the Fitbenchmark data format.
 
 from __future__ import absolute_import, division, print_function
 
-import os
-from collections import OrderedDict
 import importlib
 import inspect
+import os
+from collections import OrderedDict
 
 import numpy as np
-
 from fitbenchmarking.parsing.base_parser import Parser
 from fitbenchmarking.parsing.fitting_problem import FittingProblem
 from fitbenchmarking.utils.exceptions import MissingSoftwareError, ParsingError
@@ -34,9 +33,9 @@ except ImportError as ex:
 
 
 try:
-    from sasmodels.data import empty_data1D
-    from sasmodels.core import load_model
     from sasmodels.bumps_model import Experiment, Model
+    from sasmodels.core import load_model
+    from sasmodels.data import empty_data1D
     import_success['sasview'] = (True, None)
 except ImportError as ex:
     import_success['sasview'] = (False, ex)
@@ -100,6 +99,24 @@ class FitbenchmarkParser(Parser):
             fitting_problem.function = self._create_ivp_function()
             fitting_problem.format = 'ivp'
 
+        # If using a multivariate function wrap the call to take a single
+        # argument
+        if len(data_points[0]['x']) > 1:
+            old_function = fitting_problem.function
+            all_data = []
+            count = 0
+            for dp in data_points:
+                all_data.append(dp['x'])
+                dp['x'] = np.arange(count, count + dp['x'].shape[0])
+                count = count + dp['x'].shape[0]
+            all_data = np.concatenate(all_data)
+
+            def new_function(x, *p):
+                input = all_data[x]
+                return old_function(input, *p)
+
+            fitting_problem.function = new_function
+
         # EQUATION
         if software == 'ivp':
             fitting_problem.equation = self._equation_name
@@ -132,9 +149,9 @@ class FitbenchmarkParser(Parser):
 
         if fitting_problem.multifit:
             num_files = len(data_points)
-            fitting_problem.data_x = [d[:, 0] for d in data_points]
-            fitting_problem.data_y = [d[:, 1] for d in data_points]
-            fitting_problem.data_e = [d[:, 2] if d.shape[1] > 2 else None
+            fitting_problem.data_x = [d['x'] for d in data_points]
+            fitting_problem.data_y = [d['y'] for d in data_points]
+            fitting_problem.data_e = [d['e'] if 'e' in d else None
                                       for d in data_points]
 
             if not fit_ranges:
@@ -146,10 +163,10 @@ class FitbenchmarkParser(Parser):
                                      for f in fit_ranges]
 
         else:
-            fitting_problem.data_x = data_points[0][:, 0]
-            fitting_problem.data_y = data_points[0][:, 1]
-            if data_points[0].shape[1] > 2:
-                fitting_problem.data_e = data_points[0][:, 2]
+            fitting_problem.data_x = data_points[0]['x']
+            fitting_problem.data_y = data_points[0]['y']
+            if 'e' in data_points[0]:
+                fitting_problem.data_e = data_points[0]['e']
 
             if fit_ranges and 'x' in fit_ranges[0]:
                 fitting_problem.start_x = fit_ranges[0]['x'][0]
@@ -169,12 +186,12 @@ class FitbenchmarkParser(Parser):
 
     def _get_data_file(self):
         """
-        Find/create the (full) path to a data_file specified in a FitBenchmark
-        definition file, where the data_file is searched for in the directory
-        of the definition file and subfolders of this file.
+        Find/create the (full) path to a data_file(s) specified in a
+        FitBenchmark definition file, where the data_file is searched for in
+        the directory of the definition file and subfolders of this file.
 
[…]
