### 2021-04-26 2fdb73b
Update fitbenchmarking/systests/linux_expected_results/all_parsers.txt
Revert regression test update
