### 2021-04-27 702cc13
set constraints differently for composite functions
