### 2021-04-27 7c63cf4
compare value_ranges to lower case param names
