### 2021-04-27 e7d6948
Issue #494: Add first data assimilation example
