+++ examples/benchmark_problems/Data_Assimilation/lorentz.txt
@@ -1,0 +1,6 @@
+# Fitbenchmark Problem
+software = 'ivp'
+name = 'Lorentz'
+description = 'A simple lorentz system for testing the ivp parser.'
+input_file = 'lorentz3d.txt'
+function = 'module=functions/lorentz,func=lorentz3d,step=0.1,sigma=11,r=30,b=3'
