### 2021-04-27 380cb32
754 mantid support for bounds (#758)
