### 2021-04-28 da0832d
Better handling of failures in the table generation tests
