+++ README.md
@@ -1,6 +1,6 @@
-[![Build Status](https://img.shields.io/github/workflow/status/fitbenchmarking/fitbenchmarking/Build%20and%20Publish/v0.1.1?style=flat-square)](https://github.com/fitbenchmarking/fitbenchmarking/actions/workflows/release.yml?query=branch%3Av0.1.1)
-[![Tests Status](https://img.shields.io/github/workflow/status/fitbenchmarking/fitbenchmarking/Tests/v0.1.1?label=tests&style=flat-square)](https://github.com/fitbenchmarking/fitbenchmarking/actions/workflows/main.yml?query=branch%3Av0.1.1)
-[![Documentation Status](https://img.shields.io/readthedocs/fitbenchmarking/v0.1.1?style=flat-square)](https://fitbenchmarking.readthedocs.io/en/v0.1.1)
+[![Build Status](https://img.shields.io/github/workflow/status/fitbenchmarking/fitbenchmarking/Build%20and%20Publish?style=flat-square)](https://github.com/fitbenchmarking/fitbenchmarking/actions/workflows/release.yml)
+[![Tests Status](https://img.shields.io/github/workflow/status/fitbenchmarking/fitbenchmarking/Tests?label=tests&style=flat-square)](https://github.com/fitbenchmarking/fitbenchmarking/actions/workflows/main.yml)
+[![Documentation Status](https://img.shields.io/readthedocs/fitbenchmarking?style=flat-square)](https://fitbenchmarking.readthedocs.io/en/latest)
 [![Coverage Status](https://img.shields.io/coveralls/github/fitbenchmarking/fitbenchmarking.svg?style=flat-square)](https://coveralls.io/github/fitbenchmarking/fitbenchmarking)
 ![Windows Supported](https://img.shields.io/badge/win10-support-blue.svg?style=flat-square&logo=windows)
 ![Ubuntu Supported](https://img.shields.io/badge/18.04-support-orange.svg?style=flat-square&logo=ubuntu)
