### 2021-04-28 315bada
Issue #767: Fix linting complaints
