+++ fitbenchmarking/core/fitting_benchmarking.py
@@ -6,22 +6,24 @@ for a certain fitting software.
 
 from __future__ import absolute_import, division, print_function
 
-from collections import defaultdict
+import os
 import timeit
 import warnings
-import os
+from collections import defaultdict
 
 import numpy as np
 
 from fitbenchmarking.controllers.controller_factory import ControllerFactory
 from fitbenchmarking.cost_func.cost_func_factory import create_cost_func
-from fitbenchmarking.parsing.parser_factory import parse_problem_file
-from fitbenchmarking.utils.exceptions import UnknownMinimizerError, \
-    UnsupportedMinimizerError, ControllerAttributeError, NoJacobianError, \
-    IncompatibleMinimizerError
 from fitbenchmarking.jacobian.jacobian_factory import create_jacobian
+from fitbenchmarking.parsing.parser_factory import parse_problem_file
 from fitbenchmarking.utils import fitbm_result, misc, output_grabber
-
+from fitbenchmarking.utils.exceptions import (FitBenchmarkException,
+                                              ControllerAttributeError,
+                                              IncompatibleMinimizerError,
+                                              NoJacobianError,
+                                              UnknownMinimizerError,
+                                              UnsupportedMinimizerError)
 from fitbenchmarking.utils.log import get_logger
 
 LOGGER = get_logger()
@@ -97,17 +99,17 @@ def loop_over_benchmark_problems(problem_group, options):
         problem_passed = True
         info_str = " Running data from: {} {}/{}".format(
             os.path.basename(p), i + 1, len(problem_group))
-        LOGGER.info('\n' + '#' * (len(info_str) + 1))
+        LOGGER.info('\n%s', '#' * (len(info_str) + 1))
         LOGGER.info(info_str)
         LOGGER.info('#' * (len(info_str) + 1))
         try:
             with grabbed_output:
                 parsed_problem = parse_problem_file(p, options)
                 parsed_problem.correct_data()
-        except Exception as e:
+        except FitBenchmarkException as e:
             info_str = " Could not run data from: {} {}/{}".format(
                 p, i + 1, len(problem_group))
-            LOGGER.warn(e)
+            LOGGER.warning(e)
             problem_passed = False
 
         if problem_passed:
@@ -149,8 +151,7 @@ def loop_over_starting_values(cost_func, options, grabbed_output):
     num_start_vals = len(problem.starting_values)
     problem_results = []
     for index in range(num_start_vals):
-        LOGGER.info("    Starting value: {0}/{1}".format(index + 1,
-                                                         num_start_vals))
+        LOGGER.info("    Starting value: %i/%i", index + 1, num_start_vals)
         if num_start_vals > 1:
             problem.name = name + ', Start {}'.format(index + 1)
 
@@ -290,15 +291,15 @@ def loop_over_minimizers(controller, minimizers, options, grabbed_output):
                 minimizer_check = False
                 controller.flag = 4
                 dummy_results = [{'options': options,
-                                 'cost_func': controller.cost_func,
-                                 'jac': None,
-                                 'chi_sq': np.inf,
-                                 'runtime': np.inf,
-                                 'minimizer': minimizer,
-                                 'initial_params': controller.initial_params,
-                                 'params': None,
-                                 'error_flag': controller.flag,
-                                 'name': problem.name}]
+                                  'cost_func': controller.cost_func,
+                                  'jac': None,
+                                  'chi_sq': np.inf,
+                                  'runtime': np.inf,
+                                  'minimizer': minimizer,
+                                  'initial_params': controller.initial_params,
+                                  'params': None,
+                                  'error_flag': controller.flag,
+                                  'name': problem.name}]
                 for result in dummy_results:
                     individual_result = fitbm_result.FittingResult(
                         **result)
[…]
