### 2021-04-29 686cb89
Issue #767: Fix test issue with wrong default
