### 2021-04-30 20f48f6
remove dfo and minuit from regression tests
