+++ fitbenchmarking/systests/test_regression.py
@@ -212,11 +212,11 @@ def setup_options(multifit=False):
         opts.software = ['mantid']
         opts.minimizers = {'mantid': [opts.minimizers['mantid'][0]]}
     elif TEST_TYPE != "default":
-        opts.software = ['bumps', 'dfo', 'gsl', 'mantid', 'minuit',
+        opts.software = ['bumps', 'gsl', 'mantid', 
                          'ralfit', 'scipy', 'scipy_ls']
         opts.minimizers = {k: [v[0]] for k, v in opts.minimizers.items()}
     else:
-        opts.software = ['bumps', 'dfo', 'minuit', 'scipy', 'scipy_ls']
+        opts.software = ['bumps', 'scipy', 'scipy_ls']
         opts.minimizers = {s: [opts.minimizers[s][0]] for s in opts.software}
 
     opts.results_dir = os.path.join(os.path.dirname(__file__),
