+++ fitbenchmarking/systests/linux_expected_results/all_parsers.txt
@@ -1,8 +1,8 @@
-                                bumps                    dfo                   gsl                 mantid                 minuit                 ralfit                     scipy              scipy_ls
-                               amoeba                  dfogn                lmsder                   BFGS                 minuit                     gn               Nelder-Mead       lm-scipy-no-jac
-BENNETT5            1.608e-05 (1.001)           inf (inf)[3]  1.639e-05 (1.021)[1]      0.02038 (1269)[2]      2.114e-05 (1.316)          1.606e-05 (1)      1.653e-05 (1.029)[1]  1.904e-05 (1.186)[1]
-cubic, Start 1  1.119e-14 (2.899e+07)  5.244e-14 (1.358e+08)         3.861e-22 (1)   1.85e-12 (4.792e+09)  3.586e-11 (9.288e+10)  6.723e-13 (1.741e+09)  6.267e-05 (1.623e+17)[1]         3.861e-22 (1)
-cubic, Start 2  1.146e-14 (2.969e+07)  2.424e-17 (6.278e+04)         3.861e-22 (1)  3.306e-06 (8.563e+15)  7.579e-06 (1.963e+16)  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
-cubic-fba       1.146e-14 (2.969e+07)       7.913e-19 (2049)         3.861e-22 (1)  3.306e-06 (8.563e+15)  7.579e-06 (1.963e+16)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
-EMU 73673                    1046 (1)               1046 (1)       1053 (1.006)[1]        1553 (1.484)[2]            1057 (1.01)               1046 (1)              1055 (1.009)              1046 (1)
-Problem Def 1     0.03175 (2.445e+08)          1.299e-10 (1)   0.03966 (3.055e+08)   33.36 (2.569e+11)[2]       4.214e-07 (3245)     0.1119 (8.618e+08)      3.837e-09 (29.55)[1]   0.03966 (3.055e+08)
+                                bumps                   gsl                 mantid                 ralfit                     scipy              scipy_ls
+                               amoeba                lmsder                   BFGS                     gn               Nelder-Mead       lm-scipy-no-jac
+BENNETT5            1.608e-05 (1.001)  1.639e-05 (1.021)[1]      0.02038 (1269)[2]          1.606e-05 (1)      1.653e-05 (1.029)[1]  1.904e-05 (1.186)[1]
+cubic, Start 1  1.119e-14 (2.899e+07)         3.861e-22 (1)   1.85e-12 (4.792e+09)  6.723e-13 (1.741e+09)  6.267e-05 (1.623e+17)[1]         3.861e-22 (1)
+cubic, Start 2  1.146e-14 (2.969e+07)         3.861e-22 (1)  3.306e-06 (8.563e+15)  6.926e-18 (1.794e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
+cubic-fba       1.146e-14 (2.969e+07)         3.861e-22 (1)  3.306e-06 (8.563e+15)   9.768e-18 (2.53e+04)     7.176e-11 (1.859e+11)         3.861e-22 (1)
+EMU 73673                    1046 (1)       1053 (1.006)[1]        1553 (1.484)[2]               1046 (1)              1055 (1.009)              1046 (1)
+Problem Def 1     0.03175 (8.275e+06)   0.03966 (1.034e+07)   33.36 (8.695e+09)[2]     0.1119 (2.917e+07)          3.837e-09 (1)[1]   0.03966 (1.034e+07)
