### 2021-05-04 5cc3f0d
change rel val to inf if error flag is 5
