+++ fitbenchmarking/core/tests/test_fitting_benchmarking_jacobians.py
@@ -4,8 +4,7 @@ Tests for fitbenchmarking.core.fitting_benchmarking.loop_over_jacobians
 from __future__ import (absolute_import, division, print_function)
 import inspect
 import os
-from unittest import TestCase
-from unittest.mock import patch
+import unittest
 
 from fitbenchmarking import mock_problems
 from fitbenchmarking.cost_func.nlls_cost_func import NLLSCostFunc
@@ -18,6 +17,8 @@ from fitbenchmarking.jacobian.scipy_jacobian import Scipy
 
 # Due to construction of the controllers two folder functions
 # pylint: disable=unnecessary-pass
+
+
 class DummyController(Controller):
     """
     Minimal instantiatable subclass of Controller class for testing
@@ -85,7 +86,7 @@ def make_cost_function(file_name='cubic.dat', minimizers=None):
     return cost_func
 
 
-class LoopOverJacobiansTests(TestCase):
+class LoopOverJacobiansTests(unittest.TestCase):
     """
     loop_over_jacobians tests
     """
@@ -160,17 +161,34 @@ class LoopOverJacobiansTests(TestCase):
                    name in zip(results, new_name))
         assert new_minimizer_list == new_name
 
-    @patch.object(DummyController,"check_bounds_respected")
-    def test_bounds_respected_func_called(self, check_bounds_respected):
-        self.controller.problem.value_ranges = {'test':(0,1)}
+    @unittest.mock.patch.object(DummyController, "check_bounds_respected")
+    @unittest.mock.patch.object(DummyController, "cleanup")
+    def test_bounds_respected_func_called(
+            self, check_bounds_respected, cleanup):
+        self.controller.problem.value_ranges = {'test': (0, 1)}
         self.controller.has_jacobian = [True]
         self.controller.invalid_jacobians = ["deriv_free_algorithm"]
         self.controller.minimizer = "deriv_free_algorithm"
-        results, chi_sq, new_minimizer_list = \
-            loop_over_jacobians(self.controller,
+        self.controller.flag == 0
+        _ = loop_over_jacobians(self.controller,
                                 self.options,
                                 self.grabbed_output)
         check_bounds_respected.assert_called()
+
+    @unittest.mock.patch.object(DummyController, "check_bounds_respected")
+    @unittest.mock.patch.object(DummyController, "cleanup")
+    def test_bounds_respected_func_not_called(
+            self, check_bounds_respected, cleanup):
+        self.controller.problem.value_ranges = {'test': (0, 1)}
+        self.controller.has_jacobian = [True]
+        self.controller.invalid_jacobians = ["deriv_free_algorithm"]
+        self.controller.minimizer = "deriv_free_algorithm"
+        self.controller.flag == 3
+        _ = loop_over_jacobians(self.controller,
+                                self.options,
+                                self.grabbed_output)
+        assert not check_bounds_respected.assert_called()
+
 
 if __name__ == "__main__":
     unittest.main()
