### 2021-05-04 a0587a1
make runtine rel val inf if error flag is 5
