+++ fitbenchmarking/results_processing/compare_table.py
@@ -64,7 +64,7 @@ class CompareTable(Table):
         for key, value in results_dict.items():
             acc_abs_value = [v.chi_sq for v in value]
 
-            # exclude results with error flag 5 when finding 
+            # exclude results with error flag 5 when finding
             # min value
             error_flags = [v.error_flag for v in value]
             to_exclude = [i for i, e in enumerate(error_flags) if e == 5]
@@ -75,9 +75,15 @@ class CompareTable(Table):
             acc_rel_value = [
                 v.chi_sq / acc_min_value if v.error_flag != 5
                 else np.inf for v in value]
+
             runtime_abs_value = [v.runtime for v in value]
-            runtime_min_value = np.min(runtime_abs_value)
-            runtime_rel_value = [v.runtime / runtime_min_value for v in value]
+
+            runtime_min_value = np.min(
+                [x for i, x in enumerate(runtime_abs_value)
+                 if i not in to_exclude])
+            runtime_rel_value = [v.runtime / runtime_min_value
+                                 if v.error_flag != 5
+                                 else np.inf for v in value]
 
             abs_value[key] = [acc_abs_value, runtime_abs_value]
             rel_value[key] = [acc_rel_value, runtime_rel_value]
