+++ fitbenchmarking/jacobian/tests/test_jacobian.py
@@ -9,7 +9,7 @@ from fitbenchmarking.utils.options import Options
 from fitbenchmarking.utils import exceptions
 from fitbenchmarking.jacobian.base_jacobian import Jacobian
 from fitbenchmarking.jacobian.scipy_jacobian import Scipy
-from fitbenchmarking.jacobian.solver_jacobian import solver
+from fitbenchmarking.jacobian.minimizer_jacobian import default
 from fitbenchmarking.jacobian.numdifftools_jacobian import numdifftools
 from fitbenchmarking.jacobian.analytic_jacobian import Analytic
 from fitbenchmarking.jacobian.jacobian_factory import create_jacobian
@@ -95,12 +95,12 @@ class TestJacobianClass(TestCase):
         eval_result = jac.eval(params=self.params)
         self.assertTrue(np.isclose(self.actual, eval_result).all())
 
-    def test_solver_default(self):
+    def test_default(self):
         """
-        Test that solver default jacobian does what it should
+        Test that minimizer default jacobian does what it should
         """
-        jac = solver(self.cost_func)
-        self.assertTrue(jac.use_solver_jac)
+        jac = default(self.cost_func)
+        self.assertTrue(jac.use_default_jac)
 
     def test_numdifftools_eval(self):
         """
