### 2021-05-06 6567b5c
fix min_chi_sq
