### 2021-05-06 1c9c9ad
remove lm-scipy-no-jac from controller tests
