### 2021-05-06 a62043b
Skip plotting if using a multivariate function
