+++ fitbenchmarking/results_processing/tests/test_tables.py
@@ -187,17 +187,17 @@ class GenerateTableTests(unittest.TestCase):
             if act_line != exp_line:
                 diff.append([i, exp_line, act_line])
         if diff != []:
-            msg = "Comparing against {}".format(expected_table)) \
+            print("Comparing against {}".format(expected_table) \
                   + "\n".join(['== Line {} ==\n'\
                                'Expected :{}\n'\
                                'Actual   :{}'.format(*change)
-                               for change in diff])
+                               for change in diff]))
             print("\n==\n")
             print("Output generated (also saved as actual.out):")
             print(table)
             with open("actual.out","w") as outfile:
                 outfile.write(table)
-        self.assertListEqual([], diff, msg)
+        self.assertListEqual([], diff)
 
 
 class CreateResultsTableTests(unittest.TestCase):
