### 2021-05-07 28a9950
solver->default in docs
