+++ docs/source/users/options/jacobian_option.rst
@@ -42,13 +42,13 @@ Default is ``2-point``
     [JACOBIAN]
     scipy: 2-point
 
-.. _solver:
+.. _defaultjacobian:
 
-Solver Default Jacobian (:code:`solver`)
+Solver Default Jacobian (:code:`default`)
 --------------------------------------------
 
-This uses the approximation of the Jacobian that is used by default in the solver,
-and will vary between solvers.  If the solver requires the user to pass a Jacobian,
+This uses the approximation of the Jacobian that is used by default in the minimizer,
+and will vary between solvers.  If the minimizer requires the user to pass a Jacobian,
 a warning will be printed to the screen and the :ref:`scipy-jac` 2-point
 approximation will be used.  The only options is:
 
