+++ fitbenchmarking/cost_func/weighted_nlls_cost_func.py
@@ -1,8 +1,6 @@
 """
 Implements the weighted non-linear least squares cost function
 """
-from numpy import dot
-
 from fitbenchmarking.cost_func.nlls_base_cost_func import BaseNLLSCostFunc
 from fitbenchmarking.utils.exceptions import CostFuncError
 
