### 2021-05-10 3de7df3
change solver to default in the docs
