### 2021-05-10 bfa2954
Fix linting errors and warnings
