+++ fitbenchmarking/controllers/mantid_controller.py
@@ -45,9 +45,10 @@ class MantidController(Controller):
         else:
             raise ControllerAttributeError('Mantid Controller does not support'
                                            ' the requested cost function '
-                                           f'{self.cost_funct.__class__}')
+                                           f'{self.cost_func.__class__}')
 
-        self._param_names = self.problem.param_names
+        self.param_names = self.problem.param_names
+        self._status = None
         self.algorithm_check = {
             'all': ['BFGS', 'Conjugate gradient (Fletcher-Reeves imp.)',
                     'Conjugate gradient (Polak-Ribiere imp.)',
@@ -113,17 +114,18 @@ class MantidController(Controller):
             if self.value_ranges is not None and self._multi_fit:
                 constraints = ','.join('{0} < f{1}.{2} < {3}'.format(
                     self.value_ranges[i][0], j, p, self.value_ranges[i][1])
-                    for i, p in enumerate(self._param_names)
+                    for i, p in enumerate(self.param_names)
                     for j in range(0, self._multi_fit))
                 function_def += '; constraints=({})'.format(constraints)
             elif self.value_ranges is not None:
 
-                if not ';' in function_def:
-                    self._param_names = ['f0.'+name for name in self._param_names]
+                if ';' not in function_def:
+                    self.param_names = [
+                        'f0.'+name for name in self.param_names]
 
                 constraints = ','.join('{0} < {1} < {2}'.format(
-                        self.value_ranges[i][0], p, self.value_ranges[i][1])
-                        for i, p in enumerate(self._param_names))
+                    self.value_ranges[i][0], p, self.value_ranges[i][1])
+                    for i, p in enumerate(self.param_names))
                 function_def += '; constraints=({})'.format(constraints)
 
             self._mantid_function = function_def
@@ -153,26 +155,28 @@ class MantidController(Controller):
         if self._mantid_function is None:
             start_val_list = ['{0}={1}'.format(name, value)
                               for name, value
-                              in zip(self._param_names, self.initial_params)]
+                              in zip(self.param_names, self.initial_params)]
 
             start_val_str = ', '.join(start_val_list)
             function_def = "name=fitFunction, " + start_val_str
 
             def get_params(ff_self):
-                fit_param = np.zeros(len(self._param_names))
+                fit_param = np.zeros(len(self.param_names))
                 fit_param.setflags(write=1)
-                for i, param in enumerate(self._param_names):
+                for i, param in enumerate(self.param_names):
                     fit_param[i] = ff_self.getParameterValue(param)
                 return fit_param
 
+            # pylint: disable=E0213
             class fitFunction(IFunction1D):
                 def init(ff_self):
 
-                    for i, p in enumerate(self._param_names):
+                    for i, p in enumerate(self.param_names):
                         ff_self.declareParameter(p)
                         if self.value_ranges is not None:
                             ff_self.addConstraints('{0} < {1} < {2}'.format(
-                                self.value_ranges[i][0], p, self.value_ranges[i][1]))
+                                self.value_ranges[i][0], p,
+                                self.value_ranges[i][1]))
 
                 def function1D(ff_self, xdata):
 
@@ -185,13 +189,14 @@ class MantidController(Controller):
                     fit_param = get_params(ff_self)
 
                     jac = -self.jacobian.eval(fit_param)
-                    for i, x in enumerate(xvals):
+                    for i, _ in enumerate(xvals):
                         for j in range(len(fit_param)):
                             jacobian.set(i, j, jac[i, j])
 
             FunctionFactory.subscribe(fitFunction)
 
             self._mantid_function = function_def
+        # pylint: enable=E0213
 
     def fit(self):
         """
@@ -219,16 +224,16 @@ class MantidController(Controller):
         else:
             self.flag = 2
 
[…]
