+++ fitbenchmarking/controllers/ralfit_controller.py
@@ -26,6 +26,8 @@ class RALFitController(Controller):
         super(RALFitController, self).__init__(cost_func)
 
         self.support_for_bounds = True
+        self.param_ranges = None
+        self._status = None
         self._popt = None
         self._options = {}
         self.algorithm_check = {
