+++ fitbenchmarking/controllers/dfo_controller.py
@@ -26,6 +26,9 @@ class DFOController(Controller):
         super(DFOController, self).__init__(cost_func)
 
         self.support_for_bounds = True
+        self.param_ranges = None
+        self.rhobeg = None
+        self._status = None
         self._soln = None
         self._popt = None
         self._pinit = None
