+++ fitbenchmarking/controllers/controller_factory.py
@@ -43,11 +43,11 @@ class ControllerFactory:
             if os.path.exists(full_path):
                 raise MissingSoftwareError('Requirements are missing for the '
                                            '{} controller: {}'.format(
-                                               software, e))
+                                               software, e)) from e
             else:
                 raise NoControllerError('Could not find controller for {}. '
                                         'Check the input is correct and try '
-                                        'again.'.format(software))
+                                        'again.'.format(software)) from e
 
         classes = getmembers(module, lambda m: (isclass(m)
                                                 and not isabstract(m)
