+++ fitbenchmarking/controllers/minuit_controller.py
@@ -35,6 +35,8 @@ class MinuitController(Controller):
         super(MinuitController, self).__init__(cost_func)
 
         self.support_for_bounds = True
+        self.param_ranges = None
+        self._status = None
         self._popt = None
         self._initial_step = None
         self._minuit_problem = None
@@ -75,11 +77,11 @@ class MinuitController(Controller):
         # parameter. None is used to denote no bound for a parameter.
         if self.value_ranges is not None:
             lb, ub = zip(*self.value_ranges)
-            lb = [None if x==-np.inf else x for x in lb]
-            ub = [None if x==np.inf else x for x in ub]
+            lb = [None if x == -np.inf else x for x in lb]
+            ub = [None if x == np.inf else x for x in ub]
             self.param_ranges = list(zip(lb, ub))
         else:
-            self.param_ranges = [(-np.inf,np.inf)]*len(self.initial_params)
+            self.param_ranges = [(-np.inf, np.inf)]*len(self.initial_params)
 
         self._minuit_problem.limits = self.param_ranges
 
