+++ examples/benchmark_problems/Data_Assimilation/functions/nonlinear_oscillator.py
@@ -1,0 +1,28 @@
+import numpy as np
+
+
+def nonlinear_oscillator(t, x, gamma, mu):
+    """
+    Calculates the rhs of the system defined by:
+
+    x'[0] = x[1]
+    x'[1] = gamma*x[0]+2*mu*x[0]^3
+
+    Args:
+        t (float): The time to evaluate at
+        x (np.ndarray): The x vector (current location)
+        gamma (float): The value in the above equations
+        mu (float): The value in the above equations
+    """
+    if len(x.shape) == 1:
+        return np.array([
+            x[1],
+            gamma * x[0] + 2 * mu * x[0]**3
+        ])
+    elif len(x.shape) == 2:
+        return np.array([
+            x[:, 1],
+            gamma * x[:, 0] + 2 * mu * x[:, 0]**3
+        ])
+    else:
+        raise ValueError('x is the wrong shape in nonlinear_oscillator call.')
