### 2021-05-12 b54496d
add controller for pure matlab solution
