+++ fitbenchmarking/utils/options.py
@@ -31,6 +31,7 @@ class Options(object):
                     'Damped GaussNewton', 'Levenberg-Marquardt',
                     'Levenberg-MarquardtMD', 'Simplex',
                     'SteepestDescent', 'Trust Region'],
+         'matlab': ['Nelder-Mead Simplex'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
          'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
@@ -39,8 +40,8 @@ class Options(object):
                       'dogbox']}
     VALID_FITTING = \
         {'algorithm_type': ['all', 'ls', 'deriv_free', 'general'],
-         'software': ['bumps', 'dfo', 'gsl', 'levmar', 'mantid', 'minuit',
-                      'ralfit', 'scipy', 'scipy_ls'],
+         'software': ['bumps', 'dfo', 'gsl', 'levmar', 'mantid', 'matlab',
+                      'minuit', 'ralfit', 'scipy', 'scipy_ls'],
          'jac_method': ['scipy', 'analytic', 'numdifftools'],
          'cost_func_type': ['nlls', 'weighted_nlls', 'hellinger_nlls',
                             'poisson']}
@@ -79,6 +80,7 @@ class Options(object):
                     'Damped GaussNewton', 'Levenberg-Marquardt',
                     'Levenberg-MarquardtMD', 'Simplex',
                     'SteepestDescent', 'Trust Region'],
+         'matlab': ['Nelder-Mead Simplex'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
          'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
