### 2021-05-12 c226563
Remove cache_model_ and clear cost_func caches to avoid dill errors
