+++ fitbenchmarking/controllers/matlab_controller.py
@@ -52,6 +52,10 @@ class MatlabController(Controller):
         # Convert initial params into matlab array
         self.initial_params_mat = matlab.double([self.initial_params])
 
+        # clear out cached values
+        self.cost_func.cache_cost_x = {'params': None, 'value': None}
+        self.cost_func.cache_rx = {'params': None, 'value': None}
+
         # serialize cost_func.eval_cost and open within matlab engine
         # so that matlab fitting function can be called
         temp_dir = TemporaryDirectory()
