### 2021-05-13 4675a98
fix liniting errors/warnings in utils
