### 2021-05-13 7c70639
small fixes to controller
