### 2021-05-13 a073481
add matlab test type and add in controller tests
