+++ fitbenchmarking/controllers/tests/test_controllers.py
@@ -7,18 +7,24 @@ import pytest
 from pytest import test_type as TEST_TYPE
 
 from fitbenchmarking.controllers.base_controller import Controller
-from fitbenchmarking.controllers.bumps_controller import BumpsController
-from fitbenchmarking.controllers.controller_factory import ControllerFactory
-from fitbenchmarking.controllers.dfo_controller import DFOController
-from fitbenchmarking.controllers.minuit_controller import MinuitController
-from fitbenchmarking.controllers.scipy_controller import ScipyController
-from fitbenchmarking.controllers.scipy_ls_controller import ScipyLSController
 
-if TEST_TYPE != "default":
+if TEST_TYPE != "matlab":
+    from fitbenchmarking.controllers.bumps_controller import BumpsController
+    from fitbenchmarking.controllers.controller_factory import ControllerFactory
+    from fitbenchmarking.controllers.dfo_controller import DFOController
+    from fitbenchmarking.controllers.minuit_controller import MinuitController
+    from fitbenchmarking.controllers.scipy_controller import ScipyController
+    from fitbenchmarking.controllers.scipy_ls_controller import ScipyLSController
+
+if TEST_TYPE not in ["default", "matlab"]:
     from fitbenchmarking.controllers.gsl_controller import GSLController
     from fitbenchmarking.controllers.levmar_controller import LevmarController
     from fitbenchmarking.controllers.mantid_controller import MantidController
+    from fitbenchmarking.controllers.matlab_controller import MatlabController
     from fitbenchmarking.controllers.ralfit_controller import RALFitController
+
+if TEST_TYPE == 'matlab':
+    from fitbenchmarking.controllers.matlab_controller import MatlabController
 
 from fitbenchmarking.cost_func.weighted_nlls_cost_func import \
     WeightedNLLSCostFunc
@@ -284,7 +290,7 @@ class BaseControllerTests(TestCase):
         with self.assertRaises(exceptions.IncompatibleMinimizerError):
             controller.check_minimizer_bounds(minimizer)
 
-
+@pytest.mark.skipif("TEST_TYPE == 'matlab'")
 class DefaultControllerTests(TestCase):
     """
     Tests for each controller class
@@ -392,6 +398,7 @@ class DefaultControllerTests(TestCase):
         self.shared_tests.check_diverged(controller)
 
 @pytest.mark.skipif("TEST_TYPE == 'default'")
+@pytest.mark.skipif("TEST_TYPE == 'matlab'")
 class ControllerBoundsTests(TestCase):
 
     def setUp(self):
@@ -505,6 +512,7 @@ class ControllerBoundsTests(TestCase):
         self.check_bounds(controller)        
 
 @pytest.mark.skipif("TEST_TYPE == 'default'")
+@pytest.mark.skipif("TEST_TYPE == 'matlab'")
 class ExternalControllerTests(TestCase):
     """
     Tests for each controller class
@@ -674,6 +682,43 @@ class ExternalControllerTests(TestCase):
             controller._status = 2
             self.shared_tests.check_diverged(controller)
 
+@pytest.mark.skipif("TEST_TYPE == 'default'")
+@pytest.mark.skipif("TEST_TYPE == 'all'")
+class MatlabControllerTests(TestCase):
+    """
+    Tests for each controller class
+    """
+
+    def setUp(self):
+        self.cost_func = make_cost_func()
+        self.problem = self.cost_func.problem
+        self.jac = Scipy(self.cost_func)
+        self.jac.method = '2-point'
+        self.shared_tests = ControllerSharedTesting()
+
+    def test_matlab(self):
+        """
+        MatlabController: Tests for output shape
+        """
+        controller = MatlabController(self.cost_func)
+        controller.jacobian = self.jac
+        self.shared_tests.check_jac_info(controller,
[…]
