### 2021-05-13 a93def6
remove matlab controller from test type 'all'
