+++ fitbenchmarking/results_processing/local_min_table.py
@@ -1,8 +1,8 @@
 """
 compare table
 """
-import numpy as np
 import os
+import numpy as np
 from fitbenchmarking.results_processing.base_table import Table
 from fitbenchmarking.cost_func.nlls_base_cost_func import BaseNLLSCostFunc
 from fitbenchmarking.jacobian.jacobian_factory import create_jacobian
@@ -55,10 +55,10 @@ class LocalMinTable(Table):
         :param table_name: Name of the table
         :type table_name: str
         """
-        self.name = 'local_min'
         super(LocalMinTable, self).__init__(results, best_results, options,
                                             group_dir, pp_locations,
                                             table_name)
+        self.name = 'local_min'
 
         self.has_pp = True
         self.pp_filenames = \
@@ -103,12 +103,12 @@ class LocalMinTable(Table):
                     try:
                         jac = v.jac.eval(v.params, x=v.data_x,
                                          y=v.data_y, e=v.data_e)
-                    except:
+                    except NotImplementedError:
                         # If using a solver dependent Jacobian, we want
                         # to switch to a common jacobian here to generate
                         # comparisons
                         jacobian_cls = create_jacobian('scipy')
-                        jacobian_instance  = jacobian_cls(v.cost_func)
+                        jacobian_instance = jacobian_cls(v.cost_func)
                         jacobian_instance.method = "2-point"
                         jac = jacobian_instance.eval(v.params, x=v.data_x,
                                                      y=v.data_y, e=v.data_e)
