+++ fitbenchmarking/results_processing/plots.py
@@ -2,13 +2,13 @@
 Higher level functions that are used for plotting the fit plot and a starting
 guess plot.
 """
+import os
+import matplotlib.pyplot as plt
 import matplotlib
 matplotlib.use('Agg')
-import matplotlib.pyplot as plt
-import os
 
 
-class Plot(object):
+class Plot:
     """
     Class providing plotting functionality.
     """
@@ -58,8 +58,7 @@ class Plot(object):
         self.ax = self.fig.add_subplot(1, 1, 1)
         self.line_plot = None
 
-        use_errors = True if self.options.cost_func_type == "weighted_nlls" \
-            else False
+        use_errors = bool(self.options.cost_func_type == "weighted_nlls")
         # Plot the data that functions were fitted to
         self.plot_data(use_errors,
                        self.data_plot_options)
