+++ fitbenchmarking/controllers/scipy_controller.py
@@ -23,7 +23,7 @@ class ScipyController(Controller):
                 :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
 
         """
-        super(ScipyController, self).__init__(cost_func)
+        super().__init__(cost_func)
 
         self.support_for_bounds = True
         self.no_bounds_minimizers = ['Nelder-Mead', 'CG', 'BFGS', 'Newton-CG']
