+++ fitbenchmarking/controllers/controller_factory.py
@@ -44,10 +44,9 @@ class ControllerFactory:
                 raise MissingSoftwareError('Requirements are missing for the '
                                            '{} controller: {}'.format(
                                                software, e)) from e
-            else:
-                raise NoControllerError('Could not find controller for {}. '
-                                        'Check the input is correct and try '
-                                        'again.'.format(software)) from e
+            raise NoControllerError('Could not find controller for {}. '
+                                    'Check the input is correct and try '
+                                    'again.'.format(software)) from e
 
         classes = getmembers(module, lambda m: (isclass(m)
                                                 and not isabstract(m)
