+++ fitbenchmarking/controllers/scipy_ls_controller.py
@@ -22,7 +22,7 @@ class ScipyLSController(Controller):
         :type cost_func: subclass of
                 :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
         """
-        super(ScipyLSController, self).__init__(cost_func)
+        super().__init__(cost_func)
 
         self.support_for_bounds = True
         self.no_bounds_minimizers = ['lm-scipy']

[…]
