+++ fitbenchmarking/results_processing/acc_table.py
@@ -36,8 +36,8 @@ class AccTable(Table):
         :param table_name: Name of the table
         :type table_name: str
         """
-        super(AccTable, self).__init__(results, best_results, options,
-                                       group_dir, pp_locations, table_name)
+        super().__init__(results, best_results, options,
+                         group_dir, pp_locations, table_name)
         self.name = 'acc'
         self.has_pp = True
         self.pp_filenames = [os.path.relpath(self.pp_locations[0], group_dir)]
