+++ fitbenchmarking/controllers/base_controller.py
@@ -283,7 +283,8 @@ class Controller:
 
         The flag corresponds to the following messages:
 
-        .. automethod:: fitbenchmarking.controllers.base_controller.Controller.flag()
+        .. automethod:: fitbenchmarking.controllers.\
+                            base_controller.Controller.flag()
                 :noindex:
         """
         raise NotImplementedError
