### 2021-05-17 727e6e6
Add test for multivariate plotting handling
