+++ fitbenchmarking/cost_func/hellinger_nlls_cost_func.py
@@ -1,7 +1,7 @@
 """
 Implements the root non-linear least squares cost function
 """
-from numpy import dot, sqrt
+from numpy import dot, sqrt, ravel
 
 from fitbenchmarking.cost_func.nlls_base_cost_func import BaseNLLSCostFunc
 from fitbenchmarking.utils.exceptions import CostFuncError
@@ -60,7 +60,7 @@ class HellingerNLLSCostFunc(BaseNLLSCostFunc):
         result = sqrt(y) - sqrt(self.problem.eval_model(params=params, x=x))
 
         # Flatten in case of a vector function
-        result = result.flatten()
+        result = ravel(result)
 
         self.cache_rx['params'] = params
         self.cache_rx['value'] = result
