### 2021-05-18 92a28a4
fix linting errors and warnings in results (#799)
