+++ fitbenchmarking/results_processing/tests/test_tables.py
@@ -187,15 +187,15 @@ class GenerateTableTests(unittest.TestCase):
             if act_line != exp_line:
                 diff.append([i, exp_line, act_line])
         if diff != []:
-            print("Comparing against {}".format(expected_table) \
-                  + "\n".join(['== Line {} ==\n'\
-                               'Expected :{}\n'\
+            print("Comparing against {}".format(expected_table)
+                  + "\n".join(['== Line {} ==\n'
+                               'Expected :{}\n'
                                'Actual   :{}'.format(*change)
                                for change in diff]))
             print("\n==\n")
             print("Output generated (also saved as actual.out):")
             print(table)
-            with open("actual.out","w") as outfile:
+            with open("actual.out", "w") as outfile:
                 outfile.write(table)
         self.assertListEqual([], diff)
 
