### 2021-05-18 d1a0596
Apply suggestions from code review
