### 2021-05-18 0ab6369
Update fitbenchmarking/controllers/mantid_controller.py
