### 2021-05-18 102d185
Revert removing super args in nested scope as python cannot auto-detect the class
