### 2021-05-18 a9e8677
765 Controllers linting fixes (#790)
Co-authored-by: Andrew Lister <48282025+AndrewLister-STFC@users.noreply.github.com>
Co-authored-by: Andrew Lister <andrew.lister@stfc.ac.uk>
