+++ fitbenchmarking/controllers/scipy_controller.py
@@ -23,10 +23,13 @@ class ScipyController(Controller):
                 :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
 
         """
-        super(ScipyController, self).__init__(cost_func)
+        super().__init__(cost_func)
 
         self.support_for_bounds = True
         self.no_bounds_minimizers = ['Nelder-Mead', 'CG', 'BFGS', 'Newton-CG']
+        self.options = None
+        self.result = None
+        self._status = None
         self._popt = None
         self.algorithm_check = {
             'all': ['Nelder-Mead', 'Powell', 'CG', 'BFGS', 'Newton-CG',
