+++ fitbenchmarking/controllers/dfo_controller.py
@@ -23,9 +23,12 @@ class DFOController(Controller):
         :type cost_func: subclass of
                 :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
         """
-        super(DFOController, self).__init__(cost_func)
+        super().__init__(cost_func)
 
         self.support_for_bounds = True
+        self.param_ranges = None
+        self.rhobeg = None
+        self._status = None
         self._soln = None
         self._popt = None
         self._pinit = None
