### 2021-05-18 5577160
remove w and e from nist_parsers
