### 2021-05-19 2020781
Move cache clearing out of parser to avoid matlab clearing cache again
