### 2021-05-19 110e641
add regression tests for matlab fitting
