### 2021-05-19 02befa9
Move clear cache back to cutest parser, but only clear problems older than 1 hour
