+++ examples/benchmark_problems/Data_Assimilation/noisy_nonlinear_oscillator.txt
@@ -1,6 +1,0 @@
-# Fitbenchmark Problem
-software = 'ivp'
-name = 'Noisy Nonlinear Oscillator'
-description = 'A synthetic noisy nonlinear oscillator dataset generated to test the ivp parser. Exact results should be 10, -0.1.'
-input_file = 'noisy_nonlinear_oscillator.txt'
-function = 'module=functions/nonlinear_oscillator,func=nonlinear_oscillator,step=0.1,gamma=15,mu=-0.5'
