+++ fitbenchmarking/jacobian/numdifftools_jacobian.py
@@ -2,22 +2,15 @@
 Module which calculates numdifftools finite difference approximations
 """
 import numdifftools as nd
-import numpy as np
 
 from fitbenchmarking.jacobian.base_jacobian import Jacobian
 
-# pylint: disable=useless-super-delegation
-class numdifftools(Jacobian):
+
+class Numdifftools(Jacobian):
     """
     Implements numdifftools (https://numdifftools.readthedocs.io/en/latest/)
     finite difference approximations to the derivative
     """
-
-    def __init__(self, cost_func):
-        """
-        initializes the numdifftools jacobian
-        """
-        super(numdifftools, self).__init__(cost_func)
 
     def eval(self, params, **kwargs):
         """
@@ -29,9 +22,9 @@ class numdifftools(Jacobian):
         :return: Approximation of the Jacobian
         :rtype: numpy array
         """
-        
+
         # Use the default numdifftools derivatives
-        # 
+        #
         # Note that once jac_func is set up, it can be called
         # multiple times by giving it params.
         # We tested moving the call to nd.Jacobian into
@@ -56,6 +49,6 @@ class numdifftools(Jacobian):
         # Use the default numdifftools derivatives
         jac_cost = nd.Gradient(self.cost_func.eval_cost,
                                method=self.method)
-        
+
         jac = jac_cost(params)
         return jac
