### 2021-05-20 251e695
Increase coverage for result output
