+++ fitbenchmarking/core/tests/test_results_output.py
@@ -2,18 +2,25 @@
 Results output tests
 """
 
-from __future__ import (absolute_import, division, print_function)
-import unittest
+from __future__ import absolute_import, division, print_function
+
 import os
 import shutil
+import unittest
+from tempfile import TemporaryDirectory
+
 import numpy as np
 
+from fitbenchmarking.core.results_output import (create_directories,
+                                                 create_plots,
+                                                 create_problem_level_index,
+                                                 preproccess_data,
+                                                 save_results)
 from fitbenchmarking.cost_func.nlls_cost_func import NLLSCostFunc
 from fitbenchmarking.jacobian.scipy_jacobian import Scipy
 from fitbenchmarking.parsing.fitting_problem import FittingProblem
+from fitbenchmarking.utils.exceptions import PlottingError
 from fitbenchmarking.utils.fitbm_result import FittingResult
-from fitbenchmarking.core.results_output import create_directories, \
-    save_results, preproccess_data, create_plots, create_problem_level_index
 from fitbenchmarking.utils.options import Options
 
 
@@ -238,6 +245,8 @@ class CreatePlotsTests(unittest.TestCase):
         """
         Setting up paths and results folders
         """
+        self.tempdir = TemporaryDirectory()
+        self.dirname = os.path.join(self.tempdir.name, 'figures_dir')
         self.results, self.options, self.min_chi_sq, self.min_runtime = \
             generate_mock_results()
         self.best_results = preproccess_data(self.results)
@@ -259,7 +268,7 @@ class CreatePlotsTests(unittest.TestCase):
         # Return the above created `plot_instance`
         plot_mock.return_value = plot_instance
         create_plots(self.options, self.results,
-                     self.best_results, "figures_dir")
+                     self.best_results, self.dirname)
         for result, best_result in zip(self.results, self.best_results):
             # Check initial guess is correctly set in results
             assert best_result.start_figure_link == expected_plot_initial_guess
@@ -288,7 +297,7 @@ class CreatePlotsTests(unittest.TestCase):
         # Return the above created `plot_instance`
         plot_mock.return_value = plot_instance
         create_plots(self.options, self.results,
-                     self.best_results, "figures_dir")
+                     self.best_results, self.dirname)
         for result, best_result in zip(self.results, self.best_results):
             # Check initial guess is correctly set in results
             assert best_result.start_figure_link == expected_plot_initial_guess
@@ -306,6 +315,23 @@ class CreatePlotsTests(unittest.TestCase):
             assert best_result.figure_error == expected_message
             assert all(r.figure_error == expected_message
                        if not r.is_best_fit else True for r in result)
+
+    def test_plot_error(self):
+        """
+        Test that errors are passed correctly when the plotting fails.
+        """
+        with unittest.mock.patch(
+                'fitbenchmarking.results_processing.plots.Plot',
+                side_effect=PlottingError('Faked plot')):
+
+            create_plots(self.options, self.results,
+                         self.best_results, self.dirname)
+
+        expected = 'An error occurred during plotting.\nDetails: Faked plot'
+        for result, best_result in zip(self.results, self.best_results):
+            self.assertEqual(best_result.figure_error, expected)
+            for r in result:
+                self.assertEqual(r.figure_error, expected)
 
 
 class CreateProblemLevelIndex(unittest.TestCase):
