### 2021-05-20 4098eaf
Continue instead of return when 1 problem is multivariate on plotting
