+++ fitbenchmarking/core/results_output.py
@@ -152,7 +152,7 @@ def create_plots(options, results, best_results, figures_dir):
         except PlottingError as e:
             for result in prob_result:
                 result.figure_error = str(e)
-            return
+            continue
 
         # Create a plot showing the initial guess and get filename
         initial_guess_path = plot.plot_initial_guess()
