+++ fitbenchmarking/parsing/fitting_problem.py
@@ -184,8 +184,7 @@ class FittingProblem:
         and approximate errors if not given.
         Modifications happen on member variables.
         """
-        use_errors = True if self.options.cost_func_type == "weighted_nlls" \
-            else False
+        use_errors = self.options.cost_func_type == "weighted_nlls"
         if not self.multifit:
             correct_vals = correct_data(x=self.data_x,
                                         y=self.data_y,
@@ -236,8 +235,7 @@ class FittingProblem:
         for name in lower_param_names:
             if name in value_ranges:
                 self.value_ranges.append(
-                    (value_ranges[name][0],
-                        value_ranges[name][1]))
+                    (value_ranges[name][0], value_ranges[name][1]))
             else:
                 self.value_ranges.append((-np.inf, np.inf))
 
