### 2021-05-21 774b2e9
invert skip statements logic
