### 2021-05-24 116c1c8
Add controller for matlab optimization toolbox
