+++ fitbenchmarking/utils/options.py
@@ -31,6 +31,7 @@ class Options(object):
                     'Damped GaussNewton', 'Levenberg-Marquardt',
                     'Levenberg-MarquardtMD', 'Simplex',
                     'SteepestDescent', 'Trust Region'],
+         'matlab_opt': ['levenberg-marquardt', 'trust-region-reflective'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
          'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
@@ -38,8 +39,8 @@ class Options(object):
          'scipy_ls': ['lm-scipy', 'trf', 'dogbox']}
     VALID_FITTING = \
         {'algorithm_type': ['all', 'ls', 'deriv_free', 'general'],
-         'software': ['bumps', 'dfo', 'gsl', 'levmar', 'mantid', 'minuit',
-                      'ralfit', 'scipy', 'scipy_ls'],
+         'software': ['bumps', 'dfo', 'gsl', 'levmar', 'mantid', 'matlab_opt',
+                      'minuit', 'ralfit', 'scipy', 'scipy_ls'],
          'jac_method': ['analytic', 'scipy', 'default', 'numdifftools'],
          'cost_func_type': ['nlls', 'weighted_nlls', 'hellinger_nlls',
                             'poisson']}
@@ -79,6 +80,7 @@ class Options(object):
                     'Damped GaussNewton', 'Levenberg-Marquardt',
                     'Levenberg-MarquardtMD', 'Simplex',
                     'SteepestDescent', 'Trust Region'],
+         'matlab_opt': ['levenberg-marquardt', 'trust-region-reflective'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
          'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
