### 2021-05-24 7097cc2
Extend algorithm type
