+++ fitbenchmarking/core/fitting_benchmarking.py
@@ -288,24 +288,30 @@ def loop_over_minimizers(controller, minimizers, options, grabbed_output):
             try:
                 controller.check_minimizer_bounds(minimizer)
             except IncompatibleMinimizerError as excp:
-                minimizer_check = False
-                controller.flag = 4
-                dummy_results = [{'options': options,
-                                  'cost_func': controller.cost_func,
-                                  'jac': None,
-                                  'chi_sq': np.inf,
-                                  'runtime': np.inf,
-                                  'minimizer': minimizer,
-                                  'initial_params': controller.initial_params,
-                                  'params': None,
-                                  'error_flag': controller.flag,
-                                  'name': problem.name}]
-                for result in dummy_results:
-                    individual_result = fitbm_result.FittingResult(
-                        **result)
-                    results_problem.append(individual_result)
-                new_minimizer_list.append(minimizer)
-                LOGGER.warning(str(excp))
+                if minimizer_check:
+                    minimizer_check = False
+                    controller.flag = 4
+                    dummy_results =\
+                        [{'options': options,
+                          'cost_func': controller.cost_func,
+                          'jac': None,
+                          'chi_sq': np.inf,
+                          'runtime': np.inf,
+                          'minimizer': minimizer,
+                          'initial_params': controller.initial_params,
+                          'params': None,
+                          'error_flag': controller.flag,
+                          'name': problem.name}]
+                    # record algorithm type for specified minimizer
+                    type_str = controller.record_alg_type(
+                        minimizer, options.algorithm_type)
+                    options.minimizer_alg_type[minimizer] = type_str
+                    for result in dummy_results:
+                        individual_result = fitbm_result.FittingResult(
+                            **result)
+                        results_problem.append(individual_result)
+                    new_minimizer_list.append(minimizer)
+                    LOGGER.warning(str(excp))
 
         if minimizer_check:
             ########################
@@ -441,6 +447,11 @@ def loop_over_jacobians(controller, options, grabbed_output):
                     if controller.problem.value_ranges is not None \
                             and controller.flag != 3:
                         controller.check_bounds_respected()
+
+                    # record algorithm type for specified minimizer
+                    type_str = controller.record_alg_type(
+                        minimizer, options.algorithm_type)
+                    options.minimizer_alg_type[minimizer_name] = type_str
 
                     result_args = {'options': options,
                                    'cost_func': cost_func,
