+++ docs/source/users/options/fitting_option.rst
@@ -65,6 +65,13 @@ The options are:
 * ``deriv_free`` - derivative free algorithms (these are algorithms that cannot use
   information about derivatives -- e.g., the ``Simplex`` method in ``Mantid``)
 * ``general`` - minimizers which solve a generic `min f(x)`
+* ``simplex`` - derivative free simplect based algorithms e.g. Nelder-Mead
+* ``trust_region`` - algorithms which emply a trust region approach
+* ``levenberg-marquardt`` - minimizers that use the Levenberg Marquardt algorithm
+* ``gauss_newton`` - minimizers that use the Gauss Newton algorithm
+* ``bfgs`` - minimizers that use the BFGS algorithm
+* ``conjugate_gradient`` - Conjugate Gradient algorithms
+* ``steepest_descent`` - Steepest Descent algorithms
 
 Default is ``all``
 

[…]
