### 2021-05-24 a7823bb
fixes for tables/results output tests
