+++ fitbenchmarking/results_processing/tests/test_tables.py
@@ -115,6 +115,8 @@ def generate_mock_results():
             r.support_page_link = link_in[i][j]
             r.minimizer = options.minimizers[software][j]
             results.append(r)
+            options.minimizer_alg_type[options.minimizers[software]
+                                       [j]] = 'all, ls'
         results_out.append(results)
     best = preproccess_data(results_out)
     return best, results_out, options
