### 2021-05-25 761d854
Matlab Fitting (#795)
Co-authored-by: Andrew Lister <48282025+AndrewLister-STFC@users.noreply.github.com>

Merging with lower coverage as matlab tests can't be automated on github hosted runners.
