+++ fitbenchmarking/systests/linux_expected_results/matlab.txt
@@ -1,0 +1,5 @@
+                            matlab
+               Nelder-Mead Simplex
+BENNETT5           1.66e-05 (1)[1]
+cubic, Start 1     0.003982 (1)[1]
+cubic, Start 2       7.176e-11 (1)
