### 2021-05-25 9ba95d3
apply pylint 2.8.2 suggestions
