+++ fitbenchmarking/cost_func/weighted_nlls_cost_func.py
@@ -30,7 +30,7 @@ class WeightedNLLSCostFunc(BaseNLLSCostFunc):
                 :class:`~fitbenchmarking.parsing.fitting_problem.FittingProblem`
         """
         # Problem: The problem object from parsing
-        super(WeightedNLLSCostFunc, self).__init__(problem)
+        super().__init__(problem)
         #: *dict*
         #: Container cached residual evaluation
         self.cache_rx = {'params': None, 'value': None}
