### 2021-05-26 f3a8a29
Convert to matlab type within engine to preserve expected shape
