### 2021-05-27 12a94ae
Add documentation and add matlab_opt to all_softwares.ini
