### 2021-05-27 b9d05e1
Ensure controller respects jacobian, minimizer and bounds options
