+++ fitbenchmarking/controllers/matlab_opt_controller.py
@@ -17,7 +17,7 @@ from fitbenchmarking.utils.exceptions import CostFuncError
 eng = matlab.engine.start_matlab()
 
 
-class MatlabController(Controller):
+class MatlabOptController(Controller):
     """
     Controller for MATLAB Optimization Toolbox, implementing lsqcurvefit
     """
@@ -30,6 +30,8 @@ class MatlabController(Controller):
                 :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
         """
         super().__init__(cost_func)
+        self.support_for_bounds = True
+        self.param_ranges = None
         self.initial_params_mat = None
         self.x_data_mat = None
         self.y_data_mat = None
@@ -57,47 +59,85 @@ class MatlabController(Controller):
             raise CostFuncError('Matlab Optimization controller is not '
                                 'compatible with the chosen cost function.')
 
-
         # Convert initial params into matlab array
         self.initial_params_mat = matlab.double([self.initial_params])
         self.x_data_mat = matlab.double(self.data_x.tolist())
         self.y_data_mat = matlab.double(self.data_y.tolist())
-        
-        print('size of y data: ')
-        print(self.y_data_mat.size)
-     
-        def _eval(p, xdata):
+
+        # set matlab workspace variable for selected minimizer
+        eng.workspace['minimizer'] = self.minimizer
+
+        # set bounds if they have been set in problem definition file
+        if self.value_ranges is not None:
+            lb, ub = zip(*self.value_ranges)
+            self.param_ranges = (matlab.double(lb), matlab.double(ub))
+        else:
+            # if no bounds are set, then pass empty arrays to
+            # lsqcurvefit function
+            self.param_ranges = (matlab.double([]), matlab.double([]))
+
+        def _feval(p):
+            """
+            Function to call from matlab which evaluates the model
+            """
             feval = self.problem.eval_model(p, x=self.data_x)
             return feval
 
-        # serialize _eval and open within matlab engine
-        # so that matlab fitting function can be called
+        def _jeval(p):
+            """
+            Function to call from matlab which evaluates the jacobian
+            """
+            jeval = self.jacobian.eval(p)
+            return jeval
+
+        # serialize _feval and _jeval functions (if not using default
+        # jacobian) and open within matlab engine so
+        # matlab fitting function can be called
         temp_dir = TemporaryDirectory()
-        temp_file = os.path.join(temp_dir.name, 'temp.pickle')
-        with open(temp_file, 'wb') as f:
-            dill.dump(_eval, f)
-        eng.workspace['temp_file'] = temp_file
-        eng.evalc('py_f = py.open(temp_file,"rb")')
-        eng.evalc('eval_mat = py.dill.load(py_f)')
+        feval_file = os.path.join(temp_dir.name, 'feval.pickle')
+        with open(feval_file, 'wb') as f:
+            dill.dump(_feval, f)
+        eng.workspace['feval_file'] = feval_file
+        eng.evalc('py_f = py.open(feval_file,"rb")')
+
+        eng.evalc('eval_f = py.dill.load(py_f)')
         eng.evalc('py_f.close()')
-        eng.evalc('matlab_wrapper = @(p, x)double(eval_mat(p, x))')
+        eng.evalc('f_wrapper = @(p, x)double(eval_f(p))')
+
+        eng.workspace['eval_func'] = eng.workspace['f_wrapper']
+        eng.evalc('options = optimoptions("lsqcurvefit", \
+                                          "Algorithm", minimizer)')
+
+        # if default jacobian is selected then pass _jeval function
+        # to matlab
[…]
