### 2021-05-27 4741c17
Add in support for all nlls cost functions
