### 2021-05-27 110c273
move setting of options and eval_func into if else block
