### 2021-05-27 481551d
Cost function linting fixes (#784)
