### 2021-05-28 7bc459e
use residual, not function, in levmar
