### 2021-05-28 958f60d
move jacobian creation out of loop
