### 2021-05-28 d67a540
Update default reults directory in docs (#813)
(cherry picked from commit ef9c06b57f039463f98d1febd99ab89708621e9c)
