### 2021-05-28 46fd5b6
move jacobian object creation out from loop
