### 2021-05-28 81a2a19
Data assimilation example linting fixes
