+++ examples/benchmark_problems/Data_Assimilation/META.txt
@@ -1,4 +1,5 @@
 Data_Assimilation
 
-This a collection of simulated data assimilation problems.
-These files are not from real examples, they have been created to test the functionality within fitbenchmarking.
+These data files are synthetic and have been generated as an initial test of
+the minimizers. We plan to extend this with time series data which is more
+representative of the expectations for data assimilation in future updates.
