### 2021-05-28 58213dd
Exit pylint with exit code 1 if errors or warnings detected
