+++ requirements.txt
@@ -5,4 +5,5 @@ urllib3==1.23
 coverage<5.0
 pylint
 flake8
+pylint-exit
 pandas>=1.2
