+++ examples/benchmark_problems/SAS_modelling/SAS_1D_Modelling_Shapes_Low_Difficulty/data_files/1D_cylinder_20_400_nosmearing_neutron_synth.txt
@@ -1,0 +1,256 @@
+# <X>	<Y>	<devY>
+0.00051 414.50661 51.32677
+0.00053 435.07969 50.94081
+0.00054 495.73388 50.54420
+0.00056 425.28490 50.13667
+0.00057 422.98714 49.71790
+0.00059 479.13363 49.28759
+0.00060 394.90936 48.84543
+0.00062 461.48725 48.39108
+0.00064 434.05883 47.92422
+0.00066 502.55429 47.44451
+0.00067 456.77323 46.95159
+0.00069 409.35935 46.44511
+0.00071 503.04704 45.92468
+0.00073 408.21166 45.38995
+0.00075 500.56607 44.84051
+0.00077 424.79080 44.27597
+0.00079 447.73038 43.76830
+0.00081 356.56211 43.36155
+0.00084 452.36156 42.94357
+0.00086 445.56656 42.51405
+0.00088 409.46228 42.07269
+0.00091 434.66259 41.61916
+0.00093 499.13965 41.15313
+0.00096 448.64317 40.67427
+0.00098 431.84294 40.18223
+0.00101 440.17884 39.67666
+0.00104 512.96304 39.15719
+0.00107 404.51112 38.62345
+0.00110 428.26701 37.78976
+0.00113 445.15342 36.81138
+0.00116 461.29631 35.80644
+0.00119 472.36919 34.77426
+0.00122 419.48868 33.71413
+0.00126 471.40418 32.62535
+0.00129 509.33628 31.50718
+0.00133 479.72695 30.35888
+0.00136 440.82981 29.17971
+0.00140 395.74687 28.12530
+0.00144 466.82819 27.62008
+0.00148 427.15208 27.10113
+0.00152 429.08680 26.56810
+0.00156 526.30804 26.02063
+0.00160 482.12360 25.45836
+0.00165 465.47006 24.88093
+0.00169 427.48024 24.28798
+0.00174 426.35138 23.66542
+0.00179 452.78162 22.93541
+0.00184 452.94000 22.18601
+0.00189 443.76527 21.41680
+0.00194 456.76125 20.62730
+0.00199 412.36311 19.81708
+0.00205 434.06842 18.98568
+0.00210 431.50681 18.14134
+0.00216 487.35331 17.64398
+0.00222 448.42994 17.13356
+0.00228 452.74766 16.60979
+0.00234 452.84441 16.07241
+0.00241 433.90366 15.52113
+0.00247 425.90764 14.95569
+0.00254 431.15633 14.33554
+0.00261 446.47542 13.67547
+0.00268 443.62129 12.99886
+0.00276 431.92360 12.30544
+0.00283 435.70110 11.59494
+0.00291 443.99286 10.86710
+0.00299 421.67009 10.25550
+0.00307 433.14661 9.66868
+0.00315 431.51592 9.06785
+0.00324 427.69079 8.45284
+0.00333 437.32195 7.82351
+0.00342 433.46372 7.23946
+0.00352 418.08597 6.75555
+0.00361 428.97088 6.26069
+0.00371 424.50771 5.75484
+0.00381 422.58048 5.23794
+0.00392 421.36443 4.78241
+0.00403 418.18322 4.52129
+0.00414 418.85834 4.25453
+0.00425 417.95550 3.98216
+0.00437 417.47125 3.70418
+0.00449 415.77345 3.48361
+0.00461 403.77541 3.33285
+0.00474 412.19641 3.17898
+0.00487 412.68472 3.02203
+0.00500 404.28105 2.86205
[…]
