+++ fitbenchmarking/controllers/tests/test_controllers.py
@@ -38,8 +38,11 @@ if TEST_TYPE == 'all':
 
 if TEST_TYPE == 'matlab':
     from fitbenchmarking.controllers.matlab_controller import MatlabController
+    from fitbenchmarking.controllers.matlab_stats_controller import\
+        MatlabStatsController
 
 # pylint: disable=attribute-defined-outside-init, protected-access
+
 
 def make_cost_func(file_name='cubic.dat'):
     """
@@ -756,6 +759,26 @@ class MatlabControllerTests(TestCase):
             controller._status = 0
             self.shared_tests.check_max_iterations(controller)
             controller._status = -1
+            self.shared_tests.check_diverged(controller)
+
+    def test_matlab_stats(self):
+        """
+        MatlabStatsController: Tests for output shape
+        """
+        controller = MatlabStatsController(self.cost_func)
+        controller.jacobian = self.jac
+        self.shared_tests.check_jac_info(controller,
+                                         False,
+                                         ['Levenberg-Marquardt'])
+
+        minimizers = ['Levenberg-Marquardt']
+        for minimizer in minimizers:
+            controller.minimizer = minimizer
+            self.shared_tests.controller_run_test(controller)
+
+            controller._status = 0
+            self.shared_tests.check_converged(controller)
+            controller._status = 1
             self.shared_tests.check_diverged(controller)
 
 
