+++ fitbenchmarking/systests/test_regression.py
@@ -16,6 +16,7 @@ from conftest import run_for_test_types
 from fitbenchmarking.cli.main import run
 from fitbenchmarking.utils.options import Options
 
+
 @run_for_test_types(TEST_TYPE, 'all')
 class TestRegressionAll(TestCase):
     """
@@ -103,10 +104,12 @@ class TestRegressionAll(TestCase):
         diff, msg = diff_result(actual, expected)
         self.assertListEqual([], diff, msg)
 
+
 @run_for_test_types(TEST_TYPE, 'matlab')
 class TestRegressionMatlab(TestCase):
     """
-    Regression tests for the Fitbenchmarking software with matlab fitting software
+    Regression tests for the Fitbenchmarking software with
+    matlab fitting software
     """
 
     @classmethod
@@ -232,14 +235,14 @@ def diff_result(actual, expected):
 
     msg = '\n\nOutput has changed in {} '.format(len(diff)) \
           + 'minimizer-problem pairs. \n' \
-          + '\n'.join(['== Line {} ==\n'\
-                       'Expected :{}\n'\
+          + '\n'.join(['== Line {} ==\n'
+                       'Expected :{}\n'
                        'Actual   :{}'.format(*line_change)
                        for line_change in diff])
-    if diff !=[]:
+    if diff != []:
         print("\n==\n")
         print("Output generated (also saved as actual.out):")
-        with open("actual.out","w") as outfile:
+        with open("actual.out", "w") as outfile:
             for line in actual:
                 print(line)
                 outfile.write(line)
@@ -267,8 +270,8 @@ def setup_options(multifit=False):
                          'scipy_ls']
         opts.minimizers = {k: [v[0]] for k, v in opts.minimizers.items()}
     elif TEST_TYPE == "matlab":
-        opts.software = ['matlab']
-        opts.minimizers = {'matlab': [opts.minimizers['matlab'][0]]}
+        opts.software = ['matlab', 'matlab_stats']
+        opts.minimizers = {k: [v[0]] for k, v in opts.minimizers.items()}
     else:
         opts.software = ['bumps', 'scipy', 'scipy_ls']
         opts.minimizers = {s: [opts.minimizers[s][0]] for s in opts.software}
