+++ examples/all_softwares.ini
@@ -5,6 +5,7 @@ software: bumps
           levmar
           mantid
           matlab
+          matlab_stats
           minuit
           ralfit
           scipy

[…]
