### 2021-06-02 2c06d0c
Extend algorithm type (#809)
* Extend algorithm type

* fixes for tests

* fixes for tables/results output tests

* fix typo

* Add `gn` and `gn_reg` to levenberg marquardt tag

* Add additional tags to matlab controller

* make warning messages clearer
