+++ fitbenchmarking/utils/options.py
@@ -38,7 +38,10 @@ class Options(object):
                    'Newton-CG', 'L-BFGS-B', 'TNC', 'SLSQP'],
          'scipy_ls': ['lm-scipy', 'trf', 'dogbox']}
     VALID_FITTING = \
-        {'algorithm_type': ['all', 'ls', 'deriv_free', 'general'],
+        {'algorithm_type': ['all', 'ls', 'deriv_free', 'general', 'simplex',
+                            'trust_region', 'levenberg-marquardt',
+                            'gauss_newton', 'bfgs', 'conjugate_gradient',
+                            'steepest_descent'],
          'software': ['bumps', 'dfo', 'gsl', 'levmar', 'mantid', 'matlab',
                       'minuit', 'ralfit', 'scipy', 'scipy_ls'],
          'jac_method': ['scipy', 'analytic', 'default', 'numdifftools'],
@@ -88,7 +91,7 @@ class Options(object):
          'scipy_ls': ['lm-scipy', 'trf', 'dogbox']}
     DEFAULT_FITTING = \
         {'num_runs': 5,
-         'algorithm_type': 'all',
+         'algorithm_type': ['all'],
          'software': ['scipy', 'scipy_ls'],
          'jac_method': ['scipy'],
          'cost_func_type': 'weighted_nlls'}
@@ -163,13 +166,15 @@ class Options(object):
 
         minimizers = config['MINIMIZERS']
         self._minimizers = {}
+        self.minimizer_alg_type = {}
         for key in self.VALID_FITTING["software"]:
             self._minimizers[key] = self.read_value(minimizers.getlist,
                                                     key)
 
         fitting = config['FITTING']
         self.num_runs = self.read_value(fitting.getint, 'num_runs')
-        self.algorithm_type = self.read_value(fitting.getstr, 'algorithm_type')
+        self.algorithm_type = self.read_value(
+            fitting.getlist, 'algorithm_type')
         self.software = self.read_value(fitting.getlist, 'software')
         self.jac_method = self.read_value(fitting.getlist, 'jac_method')
         self.cost_func_type = self.read_value(fitting.getstr, 'cost_func_type')
@@ -282,7 +287,7 @@ class Options(object):
         config['MINIMIZERS'] = {k: list_to_string(m)
                                 for k, m in self.minimizers.items()}
         config['FITTING'] = {'num_runs': self.num_runs,
-                             'algorithm_type': self.algorithm_type,
+                             'algorithm_type': list_to_string(self.algorithm_type),
                              'software': list_to_string(self.software),
                              'jac_method': list_to_string(self.jac_method)}
         cs = list_to_string(['{0}, {1}'.format(*pair)
