### 2021-06-02 4ae700e
add extra algorithm type options
