### 2021-06-02 66f8309
Add extra algorithm type options & remove unused imports
