### 2021-06-02 9660cc5
investigating different sets of initial values
