+++ examples/benchmark_problems/SAS_modelling/SASView_Simple_Shapes_1D/1D_cylinder_neutron_def1.txt
@@ -1,0 +1,6 @@
+# FitBenchmark Problem
+software = 'SASView'
+name = '1D cylinder (synthetic neutron) IV#1'
+description = 'A first iteration synthetic dataset generated for the 1D cylinder SASView model in the fashion of neutron small angle scattering experiments. Generated on Fri May 28 10:31:19 2021.'
+input_file = '1D_cylinder_20_400_nosmearing_neutron_synth.txt'
+function = 'name=cylinder,radius=25.0,length=410.0,background=0.0,scale=1.0,sld=4.0,sld_solvent=1.0'
