### 2021-06-02 fe434f1
removing unneeded print statements
