### 2021-06-02 cc1d564
Add SciPy Global Optimization software
