+++ fitbenchmarking/controllers/scipy_go_controller.py
@@ -1,0 +1,93 @@
+"""
+Implements a controller for the scipy fitting software.
+In particular, here for the scipy minimize solver for general minimization
+problems.
+"""
+
+import scipy.optimize as optimize
+
+from fitbenchmarking.controllers.base_controller import Controller
+
+
+class ScipyGOController(Controller):
+    """
+    Controller for the Scipy fitting software.
+    """
+
+    def __init__(self, cost_func):
+        """
+        Initialises variable used for temporary storage.
+
+        :param cost_func: Cost function object selected from options.
+        :type cost_func: subclass of
+                :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
+
+        """
+        super(ScipyGOController, self).__init__(cost_func)
+
+        self.support_for_bounds = True
+        self._popt = None
+        self.algorithm_check = {
+            'all': ['differential_evolution', 'shgo', 'dual_annealing'],
+            'ls': [None],
+            'deriv_free': ['differential_evolution'],
+            'general': ['shgo', 'dual_annealing'],
+            'simplex': [],
+            'trust_region': [],
+            'levenberg-marquardt': [],
+            'gauss_newton': [],
+            'bfgs': [],
+            'conjugate_gradient': [],
+            'steepest_descent': [],
+            'global_optimization': ['differential_evolution', 'shgo',
+                                    'dual_annealing']
+        }
+
+    def jacobian_information(self):
+        """
+        Scipy GO can use Jacobian information
+        """
+        has_jacobian = True
+        jacobian_free_solvers = ['differential_evolution']
+        return has_jacobian, jacobian_free_solvers
+
+    def setup(self):
+        """
+        Setup problem ready to be run with SciPy GO
+        """
+        self._maxiter = 1000
+
+    def fit(self):
+        """
+        Run problem with Scipy GO.
+        """
+        if self.minimizer == "differential_evolution":
+            kwargs = {"maxiter": self._maxiter}
+        elif self.minimizer == "shgo":
+            kwargs = {"options": {"maxiter": self._maxiter,
+                                  "jac": self.jacobian.eval_cost}}
+        elif self.minimizer == "dual_annealing":
+            kwargs = {"maxiter": self._maxiter, "local_search_options": {
+                "jac": self.jacobian.eval_cost}}
+        fun = self.cost_func.eval_cost
+        if self.value_ranges is None:
+            raise RuntimeError("SciPy GO requires bounds on parameters")
+        bounds = self.value_ranges
+        algorithm = getattr(optimize, self.minimizer)
+        self.result = algorithm(fun, bounds, **kwargs)
+        self._popt = self.result.x
+        self._status = self.result.status
+
+    def cleanup(self):
+        """
+        Convert the result to a numpy array and populate the variables results
+        will be read from.
+        """
+        if self._status == 0:
+            self.flag = 0
+        elif self._status == 2:
+            self.flag = 1
+        else:
+            self.flag = 2
+
+        self.final_params = self._popt

[…]
