### 2021-06-03 0c1d8fe
Add MissingBoundsError class
