+++ fitbenchmarking/controllers/matlab_controller.py
@@ -2,20 +2,15 @@
 Implements a controller for MATLAB
 """
 
-try:
-    from tempfile import TemporaryDirectory
-except ImportError:
-    from backports.tempfile import TemporaryDirectory
-import os
-import dill
 import matlab.engine
 
-from fitbenchmarking.controllers.base_controller import Controller
+from fitbenchmarking.controllers.base_matlab_controller import\
+    BaseMatlabController
 
 eng = matlab.engine.start_matlab()
 
 
-class MatlabController(Controller):
+class MatlabController(BaseMatlabController):
     """
     Controller for MATLAB fitting (fminsearch)
     """
@@ -29,12 +24,11 @@ class MatlabController(Controller):
                 :class:`~fitbenchmarking.cost_func.base_cost_func.CostFunc`
         """
         super().__init__(cost_func)
-        self.initial_params_mat = None
         self._status = None
         self.result = None
         self.algorithm_check = {
             'all': ['Nelder-Mead Simplex'],
-            'ls': [None],
+            'ls': [],
             'deriv_free': ['Nelder-Mead Simplex'],
             'general': ['Nelder-Mead Simplex'],
             'simplex': ['Nelder-Mead Simplex'],
@@ -66,14 +60,8 @@ class MatlabController(Controller):
 
         # serialize cost_func.eval_cost and open within matlab engine
         # so that matlab fitting function can be called
-        temp_dir = TemporaryDirectory()
-        temp_file = os.path.join(temp_dir.name, 'temp.pickle')
-        with open(temp_file, 'wb') as f:
-            dill.dump(self.cost_func.eval_cost, f)
-        eng.workspace['temp_file'] = temp_file
-        eng.evalc('py_f = py.open(temp_file,"rb")')
-        eng.evalc('eval_cost_mat = py.dill.load(py_f)')
-        eng.evalc('py_f.close()')
+        eng.workspace['eval_cost_mat'] =\
+            self.py_to_mat(self.cost_func.eval_cost, eng)
 
     def fit(self):
         """
