+++ fitbenchmarking/systests/test_regression.py
@@ -154,6 +154,58 @@ class TestRegressionMatlab(TestCase):
         self.assertListEqual([], diff, msg)
 
 
+@run_for_test_types(TEST_TYPE, 'global_optimization')
+class TestRegressionGlobalOptimization(TestCase):
+    """
+    Regression tests for the Fitbenchmarking software with global optimzation
+    fitting software
+    """
+
+    @classmethod
+    def setUpClass(cls):
+        """
+        Create an options file, run it, and get the results.
+        """
+        opts = setup_options()
+        opt_file = tempfile.NamedTemporaryFile(suffix='.ini',
+                                               mode='w',
+                                               delete=False)
+        opts.write_to_stream(opt_file)
+        opt_file.close()
+        problem = os.path.abspath(os.path.join(os.path.dirname(__file__),
+                                               os.pardir,
+                                               'mock_problems',
+                                               'go_set'))
+        run([problem], options_file=opt_file.name, debug=True)
+        os.remove(opt_file.name)
+
+    def test_results_consistent_all(self):
+        """
+        Regression testing that the results of fitting a set of problems
+        containing all problem types against a single minimizer from each of
+        the supported softwares
+        """
+
+        expected_file = os.path.join(os.path.dirname(__file__),
+                                     '{}_expected_results'.format(platform),
+                                     'global_optimization.txt')
+
+        actual_file = \
+            os.path.join(os.path.dirname(__file__),
+                         'fitbenchmarking_results',
+                         'go_set',
+                         'go_set_acc_weighted_nlls_table.txt')
+
+        with open(expected_file, 'r') as f:
+            expected = f.readlines()
+
+        with open(actual_file, 'r') as f:
+            actual = f.readlines()
+
+        diff, msg = diff_result(actual, expected)
+        self.assertListEqual([], diff, msg)
+
+
 @run_for_test_types(TEST_TYPE, 'default')
 class TestRegressionDefault(TestCase):
     """
@@ -259,16 +311,20 @@ def setup_options(multifit=False):
     opts.num_runs = 1
     opts.make_plots = False
     # Use only the first minimizer from the selected software packages
+    # except for scipy_go where dual_annealing is used (others too slow) 
     if multifit:
         opts.software = ['mantid']
         opts.minimizers = {'mantid': [opts.minimizers['mantid'][0]]}
-    elif TEST_TYPE not in ['default', 'matlab']:
+    elif TEST_TYPE not in ['default', 'matlab', 'global_optimization']:
         opts.software = ['bumps', 'gsl', 'levmar', 'mantid', 'ralfit', 'scipy',
                          'scipy_ls']
         opts.minimizers = {k: [v[0]] for k, v in opts.minimizers.items()}
     elif TEST_TYPE == "matlab":
         opts.software = ['matlab']
         opts.minimizers = {'matlab': [opts.minimizers['matlab'][0]]}
+    elif TEST_TYPE == 'global_optimization':
+        opts.software = ['scipy_go']
+        opts.minimizers = {'scipy_go': ['dual_annealing']}
     else:
         opts.software = ['bumps', 'scipy', 'scipy_ls']
         opts.minimizers = {s: [opts.minimizers[s][0]] for s in opts.software}
