+++ fitbenchmarking/controllers/tests/test_controllers.py
@@ -29,8 +29,6 @@ if TEST_TYPE in ['default', 'all']:
     from fitbenchmarking.controllers.scipy_controller import ScipyController
     from fitbenchmarking.controllers.scipy_ls_controller import\
         ScipyLSController
-    from fitbenchmarking.controllers.scipy_go_controller import\
-        ScipyGOController
 
 if TEST_TYPE == 'all':
     from fitbenchmarking.controllers.gsl_controller import GSLController
@@ -40,6 +38,10 @@ if TEST_TYPE == 'all':
 
 if TEST_TYPE == 'matlab':
     from fitbenchmarking.controllers.matlab_controller import MatlabController
+
+if TEST_TYPE == 'global_optimization':
+    from fitbenchmarking.controllers.scipy_go_controller import\
+        ScipyGOController
 
 # pylint: disable=attribute-defined-outside-init, protected-access
 
@@ -451,26 +453,6 @@ class DefaultControllerTests(TestCase):
         controller._status = -1
         self.shared_tests.check_diverged(controller)
 
-    def test_scipy_go(self):
-        """
-        ScipyGOController: Test for output shape
-        """
-        controller = ScipyGOController(self.cost_func)
-        controller.minimizer = 'dual_annealing'
-        controller.jacobian = self.jac
-
-        self.shared_tests.controller_run_test(controller)
-        self.shared_tests.check_jac_info(controller,
-                                         True,
-                                         ['differential_evolution'])
-
-        controller._status = 1
-        self.shared_tests.check_converged(controller)
-        controller._status = 0
-        self.shared_tests.check_max_iterations(controller)
-        controller._status = -1
-        self.shared_tests.check_diverged(controller)
-
 
 @run_for_test_types(TEST_TYPE, 'all')
 class ControllerBoundsTests(TestCase):
@@ -805,6 +787,40 @@ class MatlabControllerTests(TestCase):
             self.shared_tests.check_max_iterations(controller)
             controller._status = -1
             self.shared_tests.check_diverged(controller)
+
+
+@run_for_test_types(TEST_TYPE, 'global_optimization')
+class GlobalOptimizationControllerTests(TestCase):
+    """
+    Tests for each controller class
+    """
+
+    def setUp(self):
+        self.cost_func = make_cost_func('cubic-fba-test-go.txt')
+        self.problem = self.cost_func.problem
+        self.jac = Scipy(self.cost_func)
+        self.jac.method = '2-point'
+        self.shared_tests = ControllerSharedTesting()
+
+    def test_scipy_go(self):
+        """
+        ScipyGOController: Test for output shape
+        """
+        controller = ScipyGOController(self.cost_func)
+        controller.minimizer = 'dual_annealing'
+        controller.jacobian = self.jac
+
+        self.shared_tests.controller_run_test(controller)
+        self.shared_tests.check_jac_info(controller,
+                                         True,
+                                         ['differential_evolution'])
+
+        controller._status = 0
+        self.shared_tests.check_converged(controller)
+        controller._status = 2
+        self.shared_tests.check_max_iterations(controller)
+        controller._status = 1
+        self.shared_tests.check_diverged(controller)
 
 
 @run_for_test_types(TEST_TYPE, 'default', 'all')
