### 2021-06-03 96a11a3
Use default test class for scipy_go
