+++ conftest.py
@@ -1,13 +1,12 @@
 import pytest
 
 # set choices for --test-type
-test_choices = ['all', 'default', 'matlab', 'global_optimization']
-
+test_choices = ['all', 'default', 'matlab']
 
 def pytest_addoption(parser):
     """
-    Command line input function which requires 'all', 'default', 'matlab'
-    or 'global_optimization' as an input
+    Command line input function which requires 'all', 'default' or 'matlab'
+    as an input
     """
     parser.addoption(
         "--test-type", action="store", default="all", type=str,
