### 2021-06-03 7c4451e
Fix regression tests
