+++ fitbenchmarking/mock_problems/go_set/META.txt
@@ -1,0 +1,3 @@
+go_set
+
+This dataset contains the global optimization test problems. It is used during testing.
