### 2021-06-04 fae4207
Try shgo instead of dual_annealing
