### 2021-06-04 ab7261d
Remove scipy_go regression tests (solvers not deterministic)
