### 2021-06-04 b7e6048
Only run controller tests for test type all
Co-authored-by: jess-farmer <79837359+jess-farmer@users.noreply.github.com>
