### 2021-06-05 #3127 @djmitche
Looks like the release was 0.1.4, on zenodo.  Looks good as far as I know -- but I'm just the reviewer, not the editor :)
