### 2021-06-07 b057b64
changing cell key for compare and local_min tables
