### 2021-06-07 22d1bcf
fix warning with newer version of pylint
