### 2021-06-07 f4f6d7c
fix errors with new version of pylint
