### 2021-06-08 f308035
Account for slow shgo solver
