### 2021-06-08 #3127 @tyronerees
@kthyng -- the version that was first reviewed was 0.1.0.  The version with the changes to the paper requested and the zenodo release (with correct metadata) is 0.1.4.  No new features added between these releases -- just a couple of bug fixes and documentation tweaks.

Please let me know if you need anything else from us.
