### 2021-06-08 7161eb4
Fix minor refactor warnings
