### 2021-06-08 59a3d4f
Fix linting in parsing (#801)
