+++ fitbenchmarking/utils/options.py
@@ -21,6 +21,18 @@ class Options(object):
     VALID_MINIMIZERS = \
         {'bumps': ['amoeba', 'lm-bumps', 'newton', 'de', 'mp'],
          'dfo': ['dfogn', 'dfols'],
+         'gradient_free' : ['HillClimbingOptimizer',
+                           'RepulsingHillClimbingOptimizer',
+                           'SimulatedAnnealingOptimizer',
+                           'RandomSearchOptimizer',
+                           'RandomRestartHillClimbingOptimizer',
+                           'RandomAnnealingOptimizer',
+                           'ParallelTemperingOptimizer',
+                           'ParticleSwarmOptimizer',
+                           'EvolutionStrategyOptimizer',
+                           'BayesianOptimizer',
+                           'TreeStructuredParzenEstimators',
+                           'DecisionTreeOptimizer'],
          'gsl': ['lmsder', 'lmder', 'nmsimplex', 'nmsimplex2',
                  'conjugate_pr', 'conjugate_fr', 'vector_bfgs',
                  'vector_bfgs2', 'steepest_descent'],
@@ -42,7 +54,7 @@ class Options(object):
                             'trust_region', 'levenberg-marquardt',
                             'gauss_newton', 'bfgs', 'conjugate_gradient',
                             'steepest_descent'],
-         'software': ['bumps', 'dfo', 'gsl', 'levmar', 'mantid', 'matlab',
+         'software': ['bumps', 'dfo', 'gradient_free', 'gsl', 'levmar', 'mantid', 'matlab',
                       'minuit', 'ralfit', 'scipy', 'scipy_ls'],
          'jac_method': ['scipy', 'analytic', 'default', 'numdifftools'],
          'cost_func_type': ['nlls', 'weighted_nlls', 'hellinger_nlls',
@@ -73,6 +85,18 @@ class Options(object):
     DEFAULT_MINIMZERS = \
         {'bumps': ['amoeba', 'lm-bumps', 'newton', 'mp'],
          'dfo': ['dfogn', 'dfols'],
+         'gradient_free': ['HillClimbingOptimizer',
+                           'RepulsingHillClimbingOptimizer',
+                           'SimulatedAnnealingOptimizer',
+                           'RandomSearchOptimizer',
+                           'RandomRestartHillClimbingOptimizer',
+                           'RandomAnnealingOptimizer',
+                           'ParallelTemperingOptimizer',
+                           'ParticleSwarmOptimizer',
+                           'EvolutionStrategyOptimizer',
+                           'BayesianOptimizer',
+                           'TreeStructuredParzenEstimators',
+                           'DecisionTreeOptimizer'],
          'gsl': ['lmsder', 'lmder', 'nmsimplex', 'nmsimplex2',
                  'conjugate_pr', 'conjugate_fr', 'vector_bfgs',
                  'vector_bfgs2', 'steepest_descent'],
