+++ docs/source/users/options/fitting_option.rst
@@ -15,6 +15,7 @@ a newline-separated list. Available options are:
 
 * ``bumps`` (default software)
 * ``dfo`` (default software)
+* ``gradient_free`` (default software)
 * ``gsl`` (external software -- see :ref:`external-instructions`)
 * ``levmar`` (external software -- see :ref:`extra_dependencies`)
 * ``mantid`` (external software -- see :ref:`external-instructions`)
@@ -25,7 +26,7 @@ a newline-separated list. Available options are:
 * ``scipy_ls`` (default software)
 
 
-Default are ``bumps``, ``dfo``, ``minuit``, ``scipy``, and ``scipy_ls``
+Default are ``bumps``, ``dfo``, ``gradient_free``, ``minuit``, ``scipy``, and ``scipy_ls``
 
 .. code-block:: rst
 
