### 2021-06-08 eddd936
Add Matlab Optimization Toolbox Fitting (#807)
* Add controller for matlab optimization toolbox

* Convert to matlab type within engine to preserve expected shape

* Add documentation and add matlab_opt to all_softwares.ini

* Ensure controller respects jacobian, minimizer and bounds options

* fix typos

* Updated list of software that support bounds

* Add in support for all nlls cost functions

* Add tests

* move setting of options and eval_func into if else block

* use residual instead of function

* fix docstrings

* Add extra algorithm type options & remove unused imports

* fix link t
[…]
