+++ fitbenchmarking/utils/options.py
@@ -32,6 +32,7 @@ class Options(object):
                     'Levenberg-MarquardtMD', 'Simplex',
                     'SteepestDescent', 'Trust Region'],
          'matlab': ['Nelder-Mead Simplex'],
+         'matlab_opt': ['levenberg-marquardt', 'trust-region-reflective'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
          'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
@@ -44,7 +45,8 @@ class Options(object):
                             'gauss_newton', 'bfgs', 'conjugate_gradient',
                             'steepest_descent', 'global_optimization'],
          'software': ['bumps', 'dfo', 'gsl', 'levmar', 'mantid', 'matlab',
-                      'minuit', 'ralfit', 'scipy', 'scipy_ls', 'scipy_go'],
+                      'matlab_opt', 'minuit', 'ralfit', 'scipy', 'scipy_ls',
+                      'scipy_go'],
          'jac_method': ['scipy', 'analytic', 'default', 'numdifftools'],
          'cost_func_type': ['nlls', 'weighted_nlls', 'hellinger_nlls',
                             'poisson']}
@@ -85,6 +87,7 @@ class Options(object):
                     'Levenberg-MarquardtMD', 'Simplex',
                     'SteepestDescent', 'Trust Region'],
          'matlab': ['Nelder-Mead Simplex'],
+         'matlab_opt': ['levenberg-marquardt', 'trust-region-reflective'],
          'minuit': ['minuit'],
          'ralfit': ['gn', 'gn_reg', 'hybrid', 'hybrid_reg'],
          'scipy': ['Nelder-Mead', 'Powell', 'CG', 'BFGS',
