### 2021-06-08 0139cc0
Add missing bounds error and tests
