### 2021-06-08 b2335c4
update docstrings/linting since we no longer pass in x
