### 2021-06-08 4522831
Edit pylint config to reduce number of tables produced
