### 2021-06-08 42c0565
X and Y labels same on log and linear plots
